% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{addmats}
\alias{addmats}
\title{adds two matrices together}
\usage{
addmats(m1, m2)
}
\arguments{
\item{m1}{matrix}

\item{m2}{matrix}
}
\value{
matrix with elements that are pointwise addition of elements in
m1 and m2
}
\description{
adds two matrices together
}
