% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent-array-to-df.R
\name{agent_array_to_df}
\alias{agent_array_to_df}
\title{Put agent_array into a data frame format}
\usage{
agent_array_to_df(agent_array)
}
\arguments{
\item{agent_array}{array of agents (n_sims x agent_id x time)}
}
\value{
data frame version of this
}
\description{
Put agent_array into a data frame format
}
