% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent-array-to-df.R
\name{agent_array_to_mat}
\alias{agent_array_to_mat}
\title{Put agent_array into a matrix format}
\usage{
agent_array_to_mat(agent_array)
}
\arguments{
\item{agent_array}{n_sims x n_agents x time array}
}
\value{
matrix with the following columns
\describe{
\item{sim}{simulation id}
\item{agent_id}{agent id}
\item{time}{time}
\item{state}{number between 0 and K-1}
}
}
\description{
Put agent_array into a matrix format
}
\details{
loop through array and make into a long matrix.  Only record when there is a change in state.  Do not record -1.  Made so to eventually replace with a Rcpp function
}
