% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{agents_sims}
\alias{agents_sims}
\title{agents_sims Example output from SIR simulations}
\format{
An array with dimension 50 x 3 x 188 data frame with 188 rows and 3 columns.  Below are the descriptions of each dimension
\describe{
  \item{sim}{Simulation number is the first dimension.  They are named between 1 and 50}
  \item{agent_stat}{This is the sufficient statistic for each agent as described above.  The names of the statistics are \code{init_state}, \code{max_time_S}, and \code{max_time_I}}
  \item{agent_id}{the ID of the agent}
}
}
\usage{
agents_sims
}
\description{
This simulated data is an example of the output from the function \code{simulate_SIR_agents()} when \code{output_format = "array"}.  It is a 50 x 3 x 188 array where entry (i,j,k) loks at the ith simulation, the jth statistic, and the kth agent.
Each column is asssociated with 1 agent and
is a "sufficient" statistic for each agent's infection.  Each agent's
infection is uniquely identified by an initial state, max time before
infection (or time T-1), and max time before recovery (or time T-1). Note time T-1
is the final recorded time in the SIR model (in this case T = 50).
}
\examples{
## show first simulation of first 10 agents
agents_sims[1, , 1:10]
}
\keyword{datasets}
