% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{agents_to_aggregate_SIR.grouped_df}
\alias{agents_to_aggregate_SIR.grouped_df}
\title{agents_to_aggregate_SIR for grouped data frames}
\usage{
\method{agents_to_aggregate_SIR}{grouped_df}(agents, max_time = NULL, ind = NULL)
}
\arguments{
\item{agents}{grouped data frame, with the following format
\describe{
  \item{init_state}{Initial state for individual (at time t = 0). For the
  states, 0 = S, 1 = I, 2 = R.}
  \item{max_time_S}{maximum time individual was susceptible (S)}
  \item{max_time_I}{maximum time individual was infected (I)}
}}

\item{max_time}{integer, max length of outbreak (default NULL), shared across
all groups}

\item{ind}{integer vector which columns match up with the columns described
above (default NULL)}
}
\value{
\code{sir_out} data frame, with columns
\describe{
  \item{grouping variable name(s)}{column/columns of grouping variable(s)}
  \item{t}{time since outbreak}
  \item{S}{Number of individuals susceptible}
  \item{I}{Number of individuals infected}
  \item{R}{Number of individuals in recovery}
}
}
\description{
agents_to_aggregate_SIR for grouped data frames
}
\examples{
library(dplyr)
max_time <- 100
agents_g <- hagelloch_raw \%>\% fortify_agents() \%>\%
  filter(SEX \%in\% c("female", "male")) \%>\% group_by(SEX)
sir_group <- agents_to_aggregate_SIR(agents_g, max_time)
agents <- agents_g \%>\%
  filter(SEX == "female") \%>\% ungroup()
sir_group1 <- agents_to_aggregate_SIR(agents, max_time)
sir_group_1 <- sir_group \%>\% filter(SEX == "female")
assertthat::are_equal(sir_group1,
                      sir_group_1 \%>\% select(t, S, I, R) \%>\% data.frame)
}
