% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{agents_to_aggregate_SIR}
\alias{agents_to_aggregate_SIR}
\title{Method for converting agent information to SIR format. Note this works
differently on \code{data.frame} and \code{grouped_df} objects.}
\usage{
agents_to_aggregate_SIR(agents, max_time = NULL, ind = NULL)
}
\arguments{
\item{agents}{data frame (grouped for ungrouped), with the following format
\describe{
  \item{init_state}{Initial state for individual (at time t = 0). For the
  states, 0 = S, 1 = I, 2 = R.}
  \item{max_time_S}{maximum time individual was susceptible (S)}
  \item{max_time_I}{maximum time individual was infected (I)}
}}

\item{max_time}{integer, max length of outbreak (default NULL)}

\item{ind}{integer vector which columns match up with the columns described
above (default NULL)}
}
\value{
\code{sir_out} data frame, with columns
\describe{
  \item{t}{time since outbreak}
  \item{S}{Number of individuals susceptible}
  \item{I}{Number of individuals infected}
  \item{R}{Number of individuals in recovery}
}
}
\description{
Method for converting agent information to SIR format. Note this works
differently on \code{data.frame} and \code{grouped_df} objects.
}
\examples{
sir_out <- agents_to_aggregate_SIR(timeternR::hagelloch_agents)
assertthat::are_equal(sir_out, timeternR::hagelloch_sir)
}
