% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-dist-mat.R
\name{as.matrix.tidy_dist_mat}
\alias{as.matrix.tidy_dist_mat}
\alias{as.array.tidy_dist_mat}
\title{convert \code{tidy_dist_mat} to \code{matrix}}
\usage{
\method{as.matrix}{tidy_dist_mat}(x)

\method{as.array}{tidy_dist_mat}(x)
}
\arguments{
\item{x}{\code{tidy_dist_mat} object}
}
\value{
matrix representation (without column information)
}
\description{
convert \code{tidy_dist_mat} to \code{matrix}
}
