% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute-recovery-counts.R
\name{cases_to_SIR.data.frame}
\alias{cases_to_SIR.data.frame}
\title{Impute Recovered counts for the SIR model}
\usage{
\method{cases_to_SIR}{data.frame}(data, par, method = "chain-binomial")
}
\arguments{
\item{data}{data frame with the following columns
\describe{
\item{t}{time}
\item{confirmed}{number of cumulative confirmed cases at time t in the group}
\item{N}{population size, which is used as the number of susceptible (minus the initial infections)}
}}

\item{par}{named vector of parameters}

\item{method}{Currently default is "chain-binomial".  See details.  More methods to come.}
}
\value{
the input data with the additional columns
\describe{
\item{X0}{number of susceptible}
\item{X1}{number of infectious}
\item{X2}{number of recovered}
}
}
\description{
Impute Recovered counts for the SIR model
}
\examples{
  df <- data.frame(t = 0:4,
                    confirmed = c(0, 1, 3, 9, 9),
                    N = 10)
   out <- cases_to_SIR(data = df,
                       par = 1)
}
