% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute-recovery-counts.R
\name{cases_to_SIR}
\alias{cases_to_SIR}
\title{Impute Recovered counts for the SIR model}
\usage{
cases_to_SIR(data, par, method = "chain-binomial")
}
\arguments{
\item{data}{data frame or grouped data frame with the following columns
\describe{
\item{t}{time}
\item{confirmed}{number of cumulative confirmed cases at time t in the group}
\item{N}{population size, which is used as the number of susceptible (minus the initial infections)}
}}

\item{par}{named vector of parameters}

\item{method}{Currently default is "chain-binomial".  See details.  More methods to come.}
}
\value{
the input data with the additional columns
\describe{
\item{X0}{number of susceptible}
\item{X1}{number of infectious}
\item{X2}{number of recovered}
}
}
\description{
Impute Recovered counts for the SIR model
}
\details{
For the method "chain-binomial". 
 Let the cumulative case counts at time \eqn{t} be \eqn{J_t}. 
  Then the number of susceptibles is simply \eqn{S_t = N - J_t}. 
   The number of infectious and recovered is imputed iteratively
    using random draws from a chain binomial based on the
     state sizes at the previous time step. 
  Specifically, we assume \eqn{I_{t_0} = J_{t_0}} and 
 \eqn{R_{t_0} = 0}, that is the initial number of recovered
  individuals is zero.  Then for each 
  \eqn{ t \in \{ t_0 + 1, t_0 + 2, \dots, T\}} 
  \eqn{R_t = R_{t-1} +} Binomial\eqn{(I_{t-1}, \gamma)} 
  and \eqn{I_t = J_t - R_t}. 
   Here \eqn{(X0, X1, X2) = (S, I, R)}.
}
\examples{
  df <- data.frame(t = 0:4,
                    confirmed = c(0, 1, 3, 9, 9),
                    N = 10)
   out <- cases_to_SIR(data = df,
                       par = 1)
}
