% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_cell_lines.R
\name{check_cell_lines}
\alias{check_cell_lines}
\title{Check cell lines}
\usage{
check_cell_lines(cell_lines = NULL, verbose = TRUE)
}
\arguments{
\item{cell_lines}{A character vector of cell line names.
If \code{NULL} (default), will return names of all cell lines.}

\item{verbose}{Print messages.}
}
\value{
Character vector, or NULL.
}
\description{
Check whether a list of cell lines matches any of those that are made
available through \code{EpiCompare}.
}
\keyword{internal}
