% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_character_percent}
\alias{check_character_percent}
\title{check if a character is a desirable percentage value}
\usage{
check_character_percent(x, name = "x")
}
\arguments{
\item{x}{character}

\item{name}{name to call \code{x} if there is an error}
}
\value{
percentage that x represents
}
\description{
check if a character is a desirable percentage value
}
