% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bands.R
\name{check_dist_params}
\alias{check_dist_params}
\title{checks the provided dist_params to make sure they are what is expected}
\usage{
check_dist_params(dist_params, data)
}
\arguments{
\item{dist_params}{named list of parameters}

\item{data}{data for visual}
}
\value{
error or updated dist_params
}
\description{
checks the provided dist_params to make sure they are what is expected
}
