% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_genome_build.R
\name{check_genome_build}
\alias{check_genome_build}
\title{Check genome build}
\usage{
check_genome_build(genome_build, type = "txdb")
}
\arguments{
\item{genome_build}{Genome build name.}

\item{type}{whether to fetch the txdb or bsgen reference data}
}
\value{
txdb or bsgen
}
\description{
Check that the genome build is valid and 
require specific reference datasets to be installed.
}
\keyword{internal}
