% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bands.R
\name{check_inside_elipsoid_func}
\alias{check_inside_elipsoid_func}
\title{create a function assert if observations are inside elipsoid}
\usage{
check_inside_elipsoid_func(Sigma, center, bound)
}
\arguments{
\item{Sigma}{covariance matrix (p x p)}

\item{center}{center of elipsoid (vector length p)}

\item{bound}{contraint for equation of ellipsoid}
}
\value{
check_inside_elipsoid_function, that takes in \code{data}
}
\description{
create a function assert if observations are inside elipsoid
}
