% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_list_names.R
\name{check_list_names}
\alias{check_list_names}
\title{Check peaklist is named}
\usage{
check_list_names(peaklist)
}
\arguments{
\item{peaklist}{A list of peak files as GRanges object.}
}
\description{
This function checks whether the peaklist is named.
If not, default file names are assigned.
}
\keyword{internal}
