% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent-to-aggregate.R
\name{check_min_max_time}
\alias{check_min_max_time}
\title{Comments for Shannon:
# not sure about how I'm dealing with the start (t_min -1 vs t_min) in example
(difference between the below examples. - this may correspond to comment:
timeternR::agents_to_aggregate_SIR.data.frame statement "## Ben, I think this
is part of the problem...") additionally, this doesn't cost us anything in
the filament visualization as the data points are the same.}
\usage{
check_min_max_time(min_max_time)
}
\arguments{
\item{min_max_time}{min_max_time}
}
\value{
error or \code{TRUE} if min_max_time vector meets assumptions
}
\description{
# example code:
agents <- timeternR::hagelloch_raw
states <- c("tI", "tR")
death <- "tDEAD"
birth <- NULL
min_max_time <- c(0, NA)
}
\examples{
b <- agents_to_aggregate(agents, states, death, birth = birth)
a <- timeternR::hagelloch_raw \%>\%
timeternR::fortify_agents() \%>\%
timeternR::agents_to_aggregate_SIR() \%>\%
as.matrix

#look at:
b[1:(nrow(b)-2),] - a[2:nrow(a),]

# keeping in mind:
timeternR::hagelloch_raw \%>\% dplyr::filter(!is.na(.data$tDEAD)) \%>\%
dplyr::pull(.data$tDEAD) \%>\% ceiling() \%>\% sort

### example 2 of code:
agents <- timeternR::hagelloch_raw
# making babies
set.seed(5)
babies <- sample(nrow(agents),size = 5)
agents$tBIRTH <- NA
agents$tBIRTH[babies] <- agents$tI[babies] - 5
states <- c("tI", "tR")
death <- NULL
birth <- "tBIRTH"
min_max_time <- c(0, NA)

b <- agents_to_aggregate(agents, states, death, birth = birth)
a <- timeternR::hagelloch_raw \%>\%
timeternR::fortify_agents() \%>\%
timeternR::agents_to_aggregate_SIR() \%>\%
as.matrix

b[1:(nrow(b)-2),] - a[2:nrow(a),]

agents \%>\% dplyr::filter(!is.na(.data$tBIRTH)) \%>\%
dplyr::pull(.data$tBIRTH) \%>\% ceiling() \%>\% sort
min_max_time vector check

this is an internal function

}
