% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thesis_data.R
\name{check_number_of_groups}
\alias{check_number_of_groups}
\title{Interior function to check number of groups from a distance matrix and cutoff}
\usage{
check_number_of_groups(dist_mat, diff_eps)
}
\arguments{
\item{dist_mat}{distance matrix}

\item{diff_eps}{threshold to suggest points are in the same group.}
}
\value{
int number of groups
}
\description{
Interior function to check number of groups from a distance matrix and cutoff
}
