% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_dist_mat.R
\name{check_tidy_dist_mat_dimensions}
\alias{check_tidy_dist_mat_dimensions}
\title{check_tidy_dist_mat_dimensions}
\usage{
check_tidy_dist_mat_dimensions(dist_mat, rownames_df, colnames_df)
}
\arguments{
\item{dist_mat}{distance matrix}

\item{rownames_df}{data.frame with row identifying information}

\item{colnames_df}{data.frame with column identifying information}
}
\value{
either a error or \code{TRUE}
}
\description{
internal function that makes sure to check the creation structure of the 
tidy_dist_mat object
}
