% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-dist-mat.R
\name{check_tidy_dist_names_distinct}
\alias{check_tidy_dist_names_distinct}
\title{check_tidy_dist_names_distinct}
\usage{
check_tidy_dist_names_distinct(rownames_df, colnames_df)
}
\arguments{
\item{rownames_df}{data.frame with row identifying information}

\item{colnames_df}{data.frame with column identifying information}
}
\value{
either a error or \code{TRUE}
}
\description{
internal function that makes sure to names are distinct
}
