% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkCache.R
\name{checkCache}
\alias{checkCache}
\title{Check cache}
\usage{
checkCache(cache = BiocFileCache::BiocFileCache(), url)
}
\arguments{
\item{cache}{BiocFileCache.}

\item{url}{Path to cached file.}
}
\value{
path
}
\description{
Quick function to check if object is already saved.
}
\keyword{internal}
