% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chromHMM_annotation_K562}
\alias{chromHMM_annotation_K562}
\title{ChromHMM annotation file}
\format{
An object of class \code{CompressedGRangesList} of length 15.
}
\source{
The code to prepare the .Rda file from the raw peak file is:
\code{
chrHMM_url <- "http://hgdownload.cse.ucsc.edu/goldenPath/hg19/encodeDCC/wgEncodeBroadHmm/wgEncodeBroadHmmK562HMM.bed.gz"
chrHMM <- genomation::readBed(chrHMM_url)
chrHMM_list <- GenomicRanges::split(chrHMM, chrHMM$name, drop = TRUE)
}
}
\usage{
chromHMM_annotation_K562
}
\description{
ChromHMM annotation for human K562 cell-line.
Obtained from \url{http://hgdownload.cse.ucsc.edu/goldenPath/hg19/encodeDCC/wgEncodeBroadHmm/}
}
\keyword{datasets}
