% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_granges.R
\name{clean_granges}
\alias{clean_granges}
\title{Clean GRanges}
\usage{
clean_granges(
  gr,
  nono_cols = c("seqnames", "ranges", "strand", "seqlevels", "seqlengths", "isCircular",
    "start", "end", "width", "element")
)
}
\arguments{
\item{gr}{A \link[GenomicRanges]{GRanges} object.}

\item{nono_cols}{Problematic columns to search for and remove (if present).}
}
\value{
Cleaned \link[GenomicRanges]{GRanges} object.
}
\description{
Remove columns from the metadata (\code{GenomicRanges::mcols}) that
conflicts with \link[GenomicRanges]{GRanges} conventions.
}
\keyword{internal}
