% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CnR_H3K27ac_picard}
\alias{CnR_H3K27ac_picard}
\title{Example Picard duplication metrics file 2}
\format{
An object of class \code{data.frame} with 1 rows and 10 columns.
}
\source{
The code to prepare the .Rda file is:

\code{
picard <- read.table("path/to/picard/duplication/output", header = TRUE, fill = TRUE)
CnR_H3K27ac_picard <- picard[1,]
usethis::use_data(CnR_H3K27ac_picard, overwrite = TRUE)
}
}
\usage{
data("CnR_H3K27ac_picard")
}
\description{
Duplication metrics output on CUT&Run H3K27ac file (sample accession: SRR8581604).
Raw sequences were aligned to hg19 genome and after, Picard was performed by Leyla Abbasova.
The duplication summary output generated by Picard was processed to reduce the size of data.
}
\keyword{datasets}
