% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CnT_H3K27ac_seacr}
\alias{CnT_H3K27ac_seacr}
\title{Example CUT&Tag SEACR called
Human H3K27ac peak file generated with CUT&Tag using K562 cell-line from Kaya-Okur et al., (2019).
Raw peak file (.BED) was obtained from GEO (sample accession: SRR8383507).
Peak calling was performed by Leyla Abbasova using SEACR.
The peak file was then processed into GRanges object.
Peaks located on chromosome 1 were subsetted to reduce the dataset size.}
\format{
An object of class \code{GRanges} of length 892.
}
\source{
THe code to prepare the .Rda file is:
\code{
CnT_H3K27ac_seacr <- ChIPseeker::readPeakFile("path/to/peak/file", as="GRanges")
CnT_H3K27ac_seacr <- CnT_H3K27ac_seacr[CnT_H3K27ac_seacr@seqnames == "chr1"]
usethis::use_data(CnT_H3K27ac_seacr)
}
}
\usage{
CnT_H3K27ac_seacr
}
\description{
Example CUT&Tag SEACR called
Human H3K27ac peak file generated with CUT&Tag using K562 cell-line from Kaya-Okur et al., (2019).
Raw peak file (.BED) was obtained from GEO (sample accession: SRR8383507).
Peak calling was performed by Leyla Abbasova using SEACR.
The peak file was then processed into GRanges object.
Peaks located on chromosome 1 were subsetted to reduce the dataset size.
}
\keyword{datasets}
