% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance-depth.R
\name{compare_new_to_rest_via_distance}
\alias{compare_new_to_rest_via_distance}
\title{Title}
\usage{
compare_new_to_rest_via_distance(
  tdmat,
  new_name_id = data.frame(id = "true observation"),
  distance_func = distance_psuedo_density_function,
  ...,
  df_out = T
)
}
\arguments{
\item{tdmat}{must be a \code{tidy_dist_mat} object}

\item{new_name_id}{data.frame containing which rownames of the \code{tdmat}
are associated with the new observations. Note that this data.frame must
be of the same format as \code{rownames(tdmat)}.}

\item{distance_func}{distance function to calculates some type of ordering
between observations with the distance matrix. This package currently
provides 3 functions for this:
\itemize{
  \item{\code{distance_psudeo_density_function}}
  \item{\code{distance_depth_function}}
  \item{\code{local_distance_depth_function}}
}
If you are trying to make your own, please see examples of these methods for
the \code{tidy_dist_mat} class. It's minimum required parameters are \code{x}
\code{x_new=NULL}, \code{df_out}. And get's passed \code{...}}

\item{...}{additional parameters to be passed to the dist_func}

\item{df_out}{indicates if one should return a data.frame or a list,
by default returns data.frame.}
}
\value{
returns either a data.frame with information about quantile score 
value for new observations, the proportion of the simulations that have more 
extreme values than it and the observations names (if df_out = T), else it 
returns a list of quatile scores and proportion of simulations that had more
extreme values than it.
}
\description{
Title
}
