% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conformal_prediction.R
\name{compression_and_sigma_estimate}
\alias{compression_and_sigma_estimate}
\title{compress filaments and estimate psuedo-density sigma in 1 go}
\usage{
compression_and_sigma_estimate(
  sim_grouped_df,
  data_columns,
  usefrac = T,
  number_points = 50,
  .sigma_string = "30\%",
  verbose = T
)
}
\arguments{
\item{sim_grouped_df}{grouped_df, set of simulated filaments}

\item{data_columns}{string of data columns that are associated with location
information}

\item{usefrac}{boolean, if the distance between filaments should scale 
relative to the number of points in compression (before sqrt)}

\item{number_points}{number of points to compressed the data with}

\item{.sigma_string}{string with percentage use for sigma (relative to 
distance matrix), e.g. \code{"30\%"}}

\item{verbose}{boolean, if we should be verbose when processing...}
}
\value{
list of compression of data as well as the sigma estimate...
}
\description{
compress filaments and estimate psuedo-density sigma in 1 go
}
