% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/containment-and-band-creation.R
\name{contained}
\alias{contained}
\alias{contained.delta_ball_structure}
\alias{contained.convex_hull_structure}
\title{Method to check if a filament is completely contained in a set (relative to
discrete representation)}
\usage{
contained(x, y)

\method{contained}{delta_ball_structure}(x, y)

\method{contained}{convex_hull_structure}(x, y)
}
\arguments{
\item{x}{assumed to either \code{convex_hull_structure} or
\code{delta_ball_structure} objects}

\item{y}{data.frame with filament represented in it}
}
\value{
boolean if y is contained in x
}
\description{
uniform containment
}
