% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conformal_prediction.R
\name{coverage_down_list}
\alias{coverage_down_list}
\title{Calculates the minimum coverage radius for filamental objects as we increase 
the number of observations.}
\usage{
coverage_down_list(data_list, e_cols, verbose = T)
}
\arguments{
\item{data_list}{list of objects ordered by correct ranking structure.}

\item{e_cols}{columns in data to compare with (in tidyverse fashion)}

\item{verbose}{boolean logic to be verbose when calculating}
}
\value{
list of min_cover_vec (minimum at that step) and dist_mat (cumulative 
max per path at time step t (column))
}
\description{
Calculates the minimum coverage radius for filamental objects as we increase 
the number of observations.
}
