% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thesis_data.R
\name{coverage_down_list_save}
\alias{coverage_down_list_save}
\title{create an array to captures the coverage_down ideas to be used for any ordering}
\usage{
coverage_down_list_save(data_list, e_cols, .e_cols_string, verbose = T)
}
\arguments{
\item{data_list}{list of data frames}

\item{e_cols}{columns associated with location of path points}

\item{.e_cols_string}{boolean if e_cols is a string (not a promise)}

\item{verbose}{boolean if we should be verbose (this can take a long time)}
}
\value{
array[i,j,k] the minimum distance from all of jth curves points to 
ith curves' kth point
}
\description{
create an array to captures the coverage_down ideas to be used for any ordering
}
