% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conformal_prediction.R
\name{coverage_down_mlist}
\alias{coverage_down_mlist}
\title{Calculates the minimum coverage radius for filamental objects as we increase 
the number of observations relative to mode groupings.}
\usage{
coverage_down_mlist(
  data_ll,
  e_cols,
  g_order_ll,
  names_df = NULL,
  .td_out = F,
  verbose = T
)
}
\arguments{
\item{data_ll}{list of objects (filaments)}

\item{e_cols}{columns in data to compare with (in tidyverse fashion)}

\item{g_order_ll}{list of indices, each vector is related to a single mode
cluster and the ordering defines how we build up the region.}

\item{names_df}{grouping information (as a data frame) asssociated with 
elements on the \code{g_list}}

\item{.td_out}{boolean, if outcome should a single "distance matrix" 
recombining different mode cluster information together. Else a list of 
"distance matrices" per mode will be returned}

\item{verbose}{boolean logic to be verbose when calculating}
}
\value{
matrix (or matrices) with the radius for each filament including at
the time step (column). See \code{.td_out} for more information.
}
\description{
Calculates the minimum coverage radius for filamental objects as we increase 
the number of observations relative to mode groupings.
}
