% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thesis_data.R
\name{coverage_down_mlist_save}
\alias{coverage_down_mlist_save}
\title{coverage_down with a saved large array matrix, multiple order vectors}
\usage{
coverage_down_mlist_save(inner_dist_mat, g_order_ll, verbose = F)
}
\arguments{
\item{inner_dist_mat}{array distance matrix, [i,j,k] is the minimum distance 
from all of jth curves points to ith curves' kth point}

\item{g_order_ll}{list of order of curves}
}
\value{
list of \code{min_cover_vec} and \code{dist_mat}
}
\description{
coverage_down with a saved large array matrix, multiple order vectors
}
