% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thesis_data.R
\name{coverage_down_slist_save}
\alias{coverage_down_slist_save}
\title{coverage_down with a saved large array matrix}
\usage{
coverage_down_slist_save(inner_dist_mat, g_order, verbose = T)
}
\arguments{
\item{inner_dist_mat}{array distance matrix, [i,j,k] is the minimum distance 
from all of jth curves points to ith curves' kth point}

\item{g_order}{vector order of curves}

\item{verbose}{bool}
}
\value{
list of \code{min_cover_vec} and \code{dist_mat}
}
\description{
coverage_down with a saved large array matrix
}
