% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conformal_prediction.R
\name{coverage_down_update}
\alias{coverage_down_update}
\title{update coverage_down information}
\usage{
coverage_down_update(cdm_list, top_percent_cut = "15\%")
}
\arguments{
\item{cdm_list}{list of lists of coverage_down information 
(\code{min_cover_vec} and \code{dist_mat})}

\item{top_percent_cut}{string with percent of the highest points to give the 
same radius as the minimum spanning value at the percent.}
}
\value{
update coverage_down information
}
\description{
update coverage_down information
}
