% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/containment-and-band-creation.R
\name{create_convex_hull_structure}
\alias{create_convex_hull_structure}
\title{create the convex hull associated with points}
\usage{
create_convex_hull_structure(
  data_points,
  data_columns = NULL,
  .lower_simplex_project = TRUE
)
}
\arguments{
\item{data_points}{points to create the convex hull structure from}

\item{data_columns}{columns of data.frame that relate to the point's
coordinates in euclidean space. This should be at least 3 columns (else it
doesn't really make sense to use this function). The input should look like
something like \code{c(S,I,R)} or \code{c("S", "I", "R")}. If the input is
\code{NULL} this function will treat this like
\code{\link[dplyr:everything]{dplyr::everything()}}.}

\item{.lower_simplex_project}{boolean, if data points should be projected to a simplex
and then to the lower dimensional simplex (for this package, this should
always be done)}
}
\value{
a \code{convex_hull_structure} object, that is a data frame similar
  to data_points (but with out \code{data_columns} columns) and only points
  that define the convex hull. This object also has an \code{normals}
  attribute which contains a matrix with the first p columns defining "A"
  and the last defining "b", such that -Ax >= b defines the convex hull.
}
\description{
I believe this will error unless one projects from the simplex to the lower
dimensional space
}
\examples{
library(dplyr)
# this will look like a 2d data frame (but this is becuase pomp_df info is
# on the 3d simplex)
convex_hull_cb <- timeternR::pomp_df \%>\%
  filter(.id <= 10) \%>\%
  arrange(time) \%>\% # just to be safe
  select(-time, -H, -cases) \%>\%
  group_by(.id) \%>\%
  grab_top_depth_filaments(alpha_level = .9) \%>\%
  create_convex_hull_structure() #data_columns = c(S,I,R)
}
