% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/containment-and-band-creation.R
\name{create_delta_ball_structure}
\alias{create_delta_ball_structure}
\title{create the delta ball associated with points}
\usage{
create_delta_ball_structure(data_points, data_columns = NULL)
}
\arguments{
\item{data_points}{points to create the delta ball structure from}

\item{data_columns}{columns (currently strings) of data.frame that relate to
the point's coordinates in euclidean space.}
}
\value{
a \code{delta_ball_structure} object, that is a data frame similar to
data_points (but with out \code{data_columns} columns) and distinct points.
This object also has an \code{delta} attribute which contains the delta
defining the delta ball.
}
\description{
create the delta ball associated with points
}
\examples{
library(dplyr)
delta_ball_cb <- timeternR::pomp_df \%>\%
  filter(.id <= 10) \%>\%
  arrange(time) \%>\% # just to be safe
  select(-time, -H, -cases) \%>\%
  group_by(.id) \%>\%
  grab_top_depth_filaments(alpha_level = .9) \%>\%
  create_delta_ball_structure()
}
