% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thesis_data.R
\name{create_discrete_uniform_cdf}
\alias{create_discrete_uniform_cdf}
\title{creates a cdf for any discrete uniform distribution (for all integer values)}
\usage{
create_discrete_uniform_cdf(min = 0, max = 1000)
}
\arguments{
\item{min}{lowest integer with non-zero prob}

\item{max}{highest integer with non-zero prob}
}
\value{
cumulative distribution function
}
\description{
creates a cdf for any discrete uniform distribution (for all integer values)
}
