% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thesis_data.R
\name{data_generation}
\alias{data_generation}
\title{generate a set of simulations from the basic SIR hierachical example based on
a specific x value.}
\usage{
data_generation(
  x_inner,
  n_sims_containment = 1000,
  number_points = 150,
  verbose = T
)
}
\arguments{
\item{x_inner}{scalar x value (between 0 -1)}

\item{n_sims_containment}{number of simulations to create}

\item{number_points}{number of points in the filamental compression}

\item{verbose}{boolean, if should report progress}
}
\value{
data frame with set of simulations after filamental compression
}
\description{
generate a set of simulations from the basic SIR hierachical example based on
a specific x value.
}
