% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta-ball.R
\name{delta_structure}
\alias{delta_structure}
\title{Performs delta ball approach}
\usage{
delta_structure(data_deep_points, xy_columns = c("x", "y"))
}
\arguments{
\item{data_deep_points}{data deep points from depth function}

\item{xy_columns}{strings for column names of the points (x,y)}
}
\value{
\describe{
\item{structure}{data frame of non-ordered lines of contour}
\item{delta}{optimal delta for covering}
}
}
\description{
Performs delta ball approach
}
