% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conformal_prediction.R
\name{diff_function_r}
\alias{diff_function_r}
\title{l2 distance between filaments}
\usage{
diff_function_r(df1, df2, usefrac = F)
}
\arguments{
\item{df1}{data.frame / matrix}

\item{df2}{data.frame / matrix (same size as m2)}

\item{usefrac}{bool if we should calculate the distance relative to a scaling
of 1/nrow(\code{df1}) (before taking the square root).}
}
\value{
numerical distance between \code{df1} and \code{df2}
}
\description{
Also not exported, similar to `difffunction `in c++
}
