% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{difffunction}
\alias{difffunction}
\title{calculates the l2 distance two matrices}
\usage{
difffunction(m1, m2, usefrac = FALSE)
}
\arguments{
\item{m1}{matrix}

\item{m2}{matrix (same size as m2)}

\item{usefrac}{bool if we should calculate the distance relative to a scaling of 
1/nrow(m1) (before taking the sqrt).}
}
\value{
numerical distance between m1 and m2
}
\description{
almost  same as EpiCompare::l2filamentdist...
}
