% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-dist-mat.R
\name{dimnames<-.tidy_dist_mat}
\alias{dimnames<-.tidy_dist_mat}
\title{assign dimnames for tidy_dist_mat}
\usage{
\method{dimnames}{tidy_dist_mat}(x) <- value
}
\arguments{
\item{x}{tidy_dist_mat}

\item{value}{list of rowname and colnames data.frame to override the 
\code{dimnames} of x with.}
}
\description{
assign dimnames for tidy_dist_mat
}
