% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-dist-mat.R
\name{dimnames.tidy_dist_mat}
\alias{dimnames.tidy_dist_mat}
\title{return dimnames for tidy_dist_mat}
\usage{
\method{dimnames}{tidy_dist_mat}(x)
}
\arguments{
\item{x}{tidy_dist_mat object}
}
\value{
list of rowname and colnames data.frames
}
\description{
return dimnames for tidy_dist_mat
}
