% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance-between.R
\name{dist_along_path}
\alias{dist_along_path}
\title{Distance/angle between points along path relative to eculidean distance}
\usage{
dist_along_path(data_df)
}
\arguments{
\item{data_df}{(n x 2) data.frame, each row is a pair of values (x,y)}
}
\value{
\item{distance}{vector of distances beween points (n - 1)}
\item{angle}{vector of angle change between points (n - 1) - in radians}
}
\description{
Calculates the distance and angle between each point of a path
}
