% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance-between.R
\name{dist_matrix_innersq_2d}
\alias{dist_matrix_innersq_2d}
\alias{dist_matrix_innersq_angle}
\title{Calculates the distance matrix between a set of paths (Euclidean based). This
is actually d^2 (2d)}
\usage{
dist_matrix_innersq_2d(path_list, position = NULL, verbose = FALSE)

dist_matrix_innersq_angle(path_list, position = NULL, verbose = FALSE)
}
\arguments{
\item{path_list}{list of paths (data frames) - need to have the same num rows}

\item{position}{column index of (x,y) euclidean coords.}

\item{verbose}{boolean logic if should have print outs while computing
distance matrix}
}
\value{
distance matrix of dimension n x n
}
\description{
Calculates the distance matrix between a set of paths (Euclidean based). This
is actually d^2 (2d)
}
\details{
This is currently built on raster::pointDistance, which only works
  in the two dimensional setting.
}
