% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance-depth.R
\name{distance_depth_function}
\alias{distance_depth_function}
\title{Geenens & Nieto-Reyes functional distance-based depth}
\usage{
distance_depth_function(dist_mat)
}
\arguments{
\item{dist_mat}{a n x n square positive symmetric matrix}
}
\value{
depth vector length n with depth values associated with indices in
  \code{dist_mat}
}
\description{
Calculates a global distance-based depth vector using a distance matrix.
Specifically we use Geenens & Nieto-Reyes's global distance-based depth
defined as:
}
\details{
\eqn{DD(x, \hat{P}) = 1/(n choose 2) * sum_{i!=j} I(d(X_1, X_2) > max(d(X_1,x),
d(X_2,x)))}


This function (renamed as \code{depth_function}) is shared with
\pkg{TCpredictionbands} on github:
\href{https://github.com/Mr8ND/TC-prediction-bands/tree/master/TCpredictionbands}{TCpredictionbands}.
}
\examples{
dist_mat <- matrix(c(0,   1, 1.5,
                     1,   0, 2,
                     1.5, 2, 0   ),
                   nrow = 3,
                   byrow = TRUE)

dd_vec <- distance_depth_function(dist_mat) # c(1,0,0)
}
