% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conformal_prediction.R
\name{distance_kernel_function_r}
\alias{distance_kernel_function_r}
\title{guassian kernel calculation relative to distances and scaling}
\usage{
distance_kernel_function_r(df, df_list, sigma)
}
\arguments{
\item{df}{filament to calculate psuedo-density estimate for}

\item{df_list}{list of filaments that define the psuedo-density}

\item{sigma}{scalar, the scaling relative to distnaces}
}
\value{
numerical value of psuedo-density estimate for \code{df}.
}
\description{
This function is an r combination of `inner_dist_density` and `distvec`
}
