% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance-psuedo-density.R
\name{distance_psuedo_density_function}
\alias{distance_psuedo_density_function}
\alias{distance_psuedo_density_function.matrix}
\alias{distance_psuedo_density_function.tidy_dist_mat}
\title{psuedo-density for objects with a distance between them}
\usage{
distance_psuedo_density_function(x, sigma = 1, df_out = "auto")

\method{distance_psuedo_density_function}{matrix}(x, sigma = 1, df_out = F)

\method{distance_psuedo_density_function}{tidy_dist_mat}(x, sigma = 1, df_out = T)
}
\arguments{
\item{x}{a n x n square positive symmetric matrix or a \code{tidy_dist_mat}}

\item{sigma}{scaling parameter. Can either by a standard numerical value or a
string as a percentage (e.g. "20\%")}

\item{df_out}{indicates if one should return a data.frame our a vector,
by default returns data.frame if dist_mat is a tidy_dist_mat, and a vector
if dist_mat is a matrix.}
}
\value{
depth vector length n with depth values associated with indices in
  \code{x} or a data.frame with a column called \code{psuedo_density}
}
\description{
DESCRIBE MORE
}
\examples{
## matrix-only examples 
dist_mat <- matrix(c(0,   1, 1.5,
                     1,   0, 2,
                     1.5, 2, 0   ),
                   nrow = 3,
                   byrow = TRUE)

dd_vec <- local_distance_depth_function(dist_mat) # c(1,0,0)

ldd_vec1 <- local_distance_depth_function(dist_mat, tau = 2) # c(1,0,0)
ldd_vec2 <- local_distance_depth_function(dist_mat, tau = 1.5) # c(1,0,0)
ldd_vec3 <- local_distance_depth_function(dist_mat, tau = 1) # c(0,0,0)
ldd_vec <- local_distance_depth_function(dist_mat, tau = .1) # c(0,0,0)
}
