% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance-between.R
\name{distanglepath_df}
\alias{distanglepath_df}
\title{Distance/direction between points along path relative to eculidean distance}
\usage{
distanglepath_df(df)
}
\arguments{
\item{df}{(n x d) data.frame, assumes rows are ordered in path}
}
\value{
\item{distance}{vector of distances beween points (n - 1)}
\item{direction}{data.frame of direction between points ((n - 1) x d)}
}
\description{
Calculates the distance and direction between each point of a
  path
}
