% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent-to-aggregate.R
\name{distinct_time}
\alias{distinct_time}
\title{select distinct rows (relative to time)}
\usage{
distinct_time(df, ignore_idx = NULL)
}
\arguments{
\item{df}{data frame, which row index related to ordering of time}

\item{ignore_idx}{index of columns of data frame that shouldn't be examined
relative to the "distinct" decision.}
}
\value{
\code{df_updated} an updated data frame with only rows with new 
information relative to time ordering
}
\description{
This function is like \code{\link[dplyr]{distinct}} but it is only relative
to the row ordering (it is also an internal function)
}
\examples{
df <- data.frame(t = 1:100, x = rep(c(1,1,2,2,3,3),
                                    length = 100),
                 y = rep(c(1,1,1,2,2,2,3,3,3),
                         length = 100))

df_filtered <- EpiCompare:::distinct_time(df, 1)
dim(df_filtered) # c(67,3)                  
df_filtered 
# same as: 
df[!(1:100 \%in\% cumsum(rep(c(2,4), length = 33))),]

}
