% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{distvec}
\alias{distvec}
\title{calculates the l2 distance between a matrix and each matrix in a list}
\usage{
distvec(m1, l2, usefrac = FALSE)
}
\arguments{
\item{m1}{primary matrix}

\item{l2}{list of matrices of the same shape of m1}

\item{usefrac}{bool if we should calculate the distance relative to a scaling of 
1/nrow(m1) (before taking the sqrt).}
}
\value{
numerical vector of distance between m1 and each matrix in list l2
}
\description{
calculates the l2 distance between a matrix and each matrix in a list
}
