% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-agents.R
\name{draws_to_states}
\alias{draws_to_states}
\title{Draw the transfer matrix}
\usage{
draws_to_states(agent_states, draw_matrix)
}
\arguments{
\item{agent_states}{vector of length agent_states where each has a value of -1, 0, ..., K-1.}

\item{draw_matrix}{KxK matrix}
}
\value{
vector of length agent_states of new states
}
\description{
Draw the transfer matrix
}
