% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiCompare.R
\name{EpiCompare}
\alias{EpiCompare}
\title{Compare epigenetic datasets}
\usage{
EpiCompare(peakfiles, file_names, blacklist, output_dir)
}
\arguments{
\item{peakfiles}{A list of peak files as GRanges object. Files must be listed using `list()`}

\item{file_names}{A list of file names in the same order as the list of peak files.
Names must be specified using `c()`}

\item{blacklist}{A GRanges object containing blacklisted regions.}

\item{output_dir}{Path to where output HTML file should be saved.}
}
\value{
An HTML report
}
\description{
Compare epigenetic datasets
}
\examples{
library(EpiCompare)
data("encode_H3K27ac") # example dataset as GRanges object
data("CnT_H3K27ac") # example dataset as GRanges object
data("hg19_blacklist") # example blacklist dataset as GRanges object

EpiCompare(peakfiles = list(encode_H3K27ac, CnT_H3K27ac),
           file_names = c("ENCODE", "CnT"),
           blacklist = hg19_blacklist,
           output_dir = "./")

}
