% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EpiModel_icm}
\alias{EpiModel_icm}
\title{EpiModel_icm Example output from the \code{EpiModel} package for a stochastic ICM}
\format{a \code{icm} object from the \code{EpiModel} package.  This contains the following elements:
\describe{
\item{param}{the epidemic parameters passed into the model through param, with additional parameters added as necessary.}
\item{control}{the control settings passed into the model through control, with additional controls added as necessary.}
\item{epi}{a list of data frames, one for each epidemiological output from the model. Outputs for base models always include the size of each compartment, as well as flows in, out of, and between compartments.}
}}
\usage{
EpiModel_icm
}
\description{
This data is made from the SIR DCM/ICM example in the \href{http://statnet.github.io/tut/BasicICMs.html}{EpiModel vignette found here}.  We run the individual contact model (ICM) for 300 steps and set the birth and death rates (a.rate, ds.rate, di.rate, dr.rate) to zero.  This is object is of class \code{icm}.  We run the simulation 10 times.  The output details can be found with \code{?EpiModel::icm}.
}
\examples{
## Look at structure
str(EpiModel_icm)

## Look at SIR values
head(EpiModel_icm$epi$s.num)
}
\keyword{datasets}
