% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bands.R
\name{equa_dist_points}
\alias{equa_dist_points}
\title{point compression}
\usage{
equa_dist_points(df, num_splits = 13)
}
\arguments{
\item{df}{2 column data frame with all points (x,y)}

\item{num_splits}{number of points for the path to be represented in}
}
\value{
new_13compression data.frame (13 x 2) of points along the path
equally spaced
}
\description{
Coverts list of point locations to a set of equally spaced points through
linear interpolation.
}
\examples{
library(tidyverse)
my_df <- data.frame(x = 1:20) \%>\%
  mutate(y = x)
my_df_compression <- equa_dist_points(my_df)
my_df_compression
}
