% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance-between.R
\name{equa_dist_points_angle}
\alias{equa_dist_points_angle}
\title{point compression (2d)}
\usage{
equa_dist_points_angle(df, num_splits = 13)
}
\arguments{
\item{df}{2 column data frame with all points (x,y)}

\item{num_splits}{integer number of points for the path to be represented in}
}
\value{
new_compression data.frame (num_splits x 2) of points along the path
equally spaced
}
\description{
Coverts list of point locations to a set of equally spaced points through
linear interpolation.
}
\examples{
library(dplyr)
my_df <- data.frame(x = 1:20) \%>\%
  mutate(y = x)
my_df_compression <- equa_dist_points_angle(my_df)
my_df_compression
}
