% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bands.R
\name{equa_dist_points_listable}
\alias{equa_dist_points_listable}
\title{List of compressions}
\usage{
equa_dist_points_listable(
  list_df,
  position = 1:2,
  num_splits = 13,
  verbose = TRUE
)
}
\arguments{
\item{list_df}{list of dfs, where the (x,y) points are in the position
columns}

\item{position}{the columns of the data frames that contain the desired
(x,y) coordinates}

\item{num_splits}{number of points for the path to be represented in}

\item{verbose}{boolean for having a progress bar}
}
\value{
list will the paths with the same length
}
\description{
Creates list of same length of compression for each path
}
