% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent-to-aggregate.R
\name{expanding_info}
\alias{expanding_info}
\title{expanding aggregate data to desirable format}
\usage{
expanding_info(df, t_min, t_max, K)
}
\arguments{
\item{df}{data frame with \code{state}, \code{t} and \code{count} columns
which provide information on the number of agents that became said
\code{state} at time \code{t}.}

\item{t_min}{minimum integer time}

\item{t_max}{maximum integer time}

\item{K}{(number of stages - 1) (e.g.: SIR has K = 2)}
}
\value{
expanded data frame: a data frame with columns \code{t} and a set of
  the unique \code{states} as columns. For each row, it contains the number
  of agents to change to said state at specified time.
}
\description{
This internal function specifically takes a data frame the contains rows that
provide the an integer time, a state and the number of agents that entered
that state at that time (this data frame only contains counts > 0).
}
\details{
This function then outputs a "\code{pivot_longer}" data frame, that has 1 row
per integer time period between (and including) t_min and t_max. Even if 0
agents change their state at that given time. The columns (beyond \code{t})
relate to each state.
}
