% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde.R
\name{extract_contour}
\alias{extract_contour}
\title{Selection of specific contour level from KDE object}
\usage{
extract_contour(kde_obj, alpha)
}
\arguments{
\item{kde_obj}{kde object}

\item{alpha}{contour level which needs to be extracted. Scalar in the
sequence from .01 to .99 by .01. A value of .95 gives a contour level
associated with the cumulative prob of 95\% (and 95\% confidence interval) -
which is technically 1-alpha.}
}
\value{
List of contours at (100*alpha level) for the kde object - there is a
contour for each disconnected contour that forms the (100*alpha level)
}
\description{
This function extracts a specific level of contour from a kde object.

This function only works for some levels that the kde object calculated
originally (associated with the constraints in the \code{alpha} parameter).
}
\details{
This function (renamed as \code{extract_countour}) is shared with \pkg{TCpredictionbands} on github:
\href{https://github.com/Mr8ND/TC-prediction-bands/tree/master/TCpredictionbands}{TCpredictionbands}.
}
\examples{
\dontrun{
set.seed(8192)

x1 <- 2^rnorm(100)
y1 <- rnorm(100)
dfmat <- cbind(x1,y1)
kde_object <- ks::kde(dfmat)

cont <- extract_contour(kde_object, .05)
}
}
