% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify-aggregate-epimodel-icm.R
\name{extract_icm_cols}
\alias{extract_icm_cols}
\title{Extract the icm class totals}
\usage{
extract_icm_cols(nm, ii, epi)
}
\arguments{
\item{nm}{name of variable to extract}

\item{ii}{integer between 0 and K - this is the state number}

\item{epi}{list of matrcies from icm output}
}
\value{
a tibble
}
\description{
Extract the icm class totals
}
