% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_length.R
\name{fig_length}
\alias{fig_length}
\title{Dynamic Figure Length Generator}
\usage{
fig_length(default_size, number_of_items, max_items)
}
\arguments{
\item{default_size}{The default figure length. Must be numeric.}

\item{number_of_items}{Number of peak files, or terms.}

\item{max_items}{Maximum number of peak files, or terms.}
}
\value{
Figure height/width. A number.
}
\description{
This function calculates the appropriate figure height depending on the
number of items.
}
\keyword{internal}
