% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/containment-and-band-creation.R
\name{filament_distance_depth}
\alias{filament_distance_depth}
\title{Calculate filament depth (relative to distance depth)}
\usage{
filament_distance_depth(
  grouped_df,
  data_columns = NULL,
  dist_list_func = "auto",
  depth_func = distance_depth_function
)
}
\arguments{
\item{grouped_df}{grouped_df data.frame object (assumed rows per filament are
ordered) - grouped per each filament}

\item{data_columns}{columns of data.frame that relate to
  the filament's coordinates in euclidean space. The input should look like
something like \code{c(S,I,R)} or \code{c("S", "I", "R")}. If the input is
\code{NULL} this function will treat this like all non-group columns.}

\item{dist_list_func}{function to calculate distance between each filament.
The default is "auto", which is selects either
\code{dist_matrix_innersq_direction} if all filaments are the same length
-or- preforms a 13 point compression of each filament using
\code{filament_compression}.}

\item{depth_func}{function to distances observed between filaments, default
is \code{distance_depth_function}}
}
\value{
depth_vector (of depth related to each filament)
}
\description{
Calculate filament depth (relative to distance depth)
}
