% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde.R
\name{fit_kde_object}
\alias{fit_kde_object}
\title{Fit kernel density estimator (KDE) object}
\usage{
fit_kde_object(df, h_band = NULL, position = 1:2, grid_size = rep(1000, 2))
}
\arguments{
\item{df}{data.frame with all the points}

\item{h_band}{optional argument for the bandwidth of the kde object. If NULL,
the optimal band would be selected through the \code{\link[ks:kde]{ks::kde}}
function. Default is NULL.}

\item{position}{Columns position of x/y pair. Default is 1:2}

\item{grid_size}{size of the grid which is going to be used for the
evaluation of kde object. Can be reduced to speed-up computation.}
}
\value{
KDE object fitted to the df data.frame.
}
\description{
This function fits a kernel density estimator to a set of data points.
Another wrapper of \code{\link[ks:kde]{ks::kde}}.
}
\details{
This function is shared with \pkg{TCpredictionbands} on github:
\href{https://github.com/Mr8ND/TC-prediction-bands/tree/master/TCpredictionbands}{TCpredictionbands}.
}
\examples{
\dontrun{
set.seed(8192)
x <- 2^rnorm(100)
y <- rnorm(100)
dfmat <- cbind(x,y)

kde_object <- fit_kde_object(dfmat)
}
}
