% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-dist-mat.R
\name{format.tidy_dist_mat}
\alias{format.tidy_dist_mat}
\title{Fomat tidy_dist_mat for printing}
\usage{
\method{format}{tidy_dist_mat}(x, ...)
}
\arguments{
\item{x}{tid_dist_mat}

\item{...}{additional parameter (e.g. \code{digits}) - tells us the number of 
significant digits to present for the distances}
}
\value{
format ready information
}
\description{
Fomat tidy_dist_mat for printing
}
