% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify_data.R
\name{fortify.dcm}
\alias{fortify.dcm}
\title{Takes in output from the \code{R} \code{EpiModel} package in the \code{dcm} class and puts it in SIR format}
\usage{
\method{fortify}{dcm}(EpiModel_output)
}
\arguments{
\item{EpiModel_output}{output from \code{EpiModel::dcm}}
}
\value{
data frame with the following columns
\describe{
\item{t}{the time}
\item{S}{number of Susceptibles}
\item{I}{number of Infectious}
\item{R}{number of Recovered}
\item{sim}{simulation number (factor variable) (optional column)}
}
}
\description{
Takes in output from the \code{R} \code{EpiModel} package in the \code{dcm} class and puts it in SIR format
}
\details{
Take the output from \code{EpiModel::dcm} and turn it into an SIR data.frame for plotting.
}
\examples{
## For dcm
sir1 <- fortify(EpiModel_det)
head(sir1)
class(sir1)
}
