% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify_data.R
\name{fortify.icm}
\alias{fortify.icm}
\title{Takes in output from the \code{R} \code{EpiModel} package in \code{icm} format and puts it in SIR format}
\usage{
\method{fortify}{icm}(data, ...)
}
\arguments{
\item{data}{output from  \code{EpiModel::icm}}

\item{...}{additional arguments}
}
\value{
data frame with the following columns
\describe{
\item{t}{the time}
\item{S}{number of Susceptibles}
\item{I}{number of Infectious}
\item{R}{number of Recovered}
\item{sim}{simulation number (factor variable) (optional column)}
}
}
\description{
Takes in output from the \code{R} \code{EpiModel} package in \code{icm} format and puts it in SIR format
}
\details{
Take the output from \code{EpiModel::icm} and turn it into an SIR data.frame for plotting.
}
\examples{
## For icm
sir <- fortify(EpiModel_icm)
head(sir)
class(sir)
}
