% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify_data.R
\name{fortify_agents}
\alias{fortify_agents}
\title{Fortify agent data frame with columns when individual stops being susceptible
and stops being infected (as well as initial state).}
\usage{
fortify_agents(
  data,
  time_col = c("tI", "tR"),
  max_time = floor(max(raw_df[, time_col], na.rm = TRUE)) + 1
)
}
\arguments{
\item{data}{data frame, agent based data frame, class ("individuals_df")}

\item{time_col}{length 2 string vector, column names recording when
individual is infected and when they enter the recovery stage}

\item{max_time}{int, maximum time for infection process}
}
\value{
\code{fortified_df} data frame, the \code{raw_df} plus three
additional columns:
\describe{
  \item{init_state}{Initial state for individual (at time t = 0). For the
  states, 0 = S, 1 = I, 2 = R.}
  \item{max_time_S}{maximum time individual was susceptible (S)}
  \item{max_time_I}{maximum time individual was infected (I)}
}
}
\description{
Fortify agent data frame with columns when individual stops being susceptible
and stops being infected (as well as initial state).
}
\examples{
fortify_df <- fortify_agents(timeternR::hagelloch_raw,
                             time_col = c("tI","tR"),
                             max_time = 95)
assertthat::are_equal(fortify_df[,c("init_state",
"max_time_S", "max_time_I")],
                      timeternR::hagelloch_agents)
}
