% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify-aggregate-epimodel-inner.R
\name{fortify_aggregate.epimodel_inner}
\alias{fortify_aggregate.epimodel_inner}
\title{Take EpiModel Data and format it}
\usage{
\method{fortify_aggregate}{epimodel_inner}(data, states, package_source = NULL)
}
\arguments{
\item{data}{output from external source package.  See details}

\item{states}{names of states we want aggregate totals of at each time.  This is a \code{string}.}

\item{package_source}{optional argument to include the package from which the
output is derived from, which helps with the fortify function when outputs
are of generic classes such as list or data.frame}
}
\value{
a data frame with the following columns
\describe{
\item{t}{time}
\item{Xk}{columns X0, ..., X_K. which are numeric}
}
}
\description{
Take EpiModel Data and format it
}
\details{
This function converts epimodel data sources and converts
  and puts it in a format that can be used by our exploring functions.
}
