% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify-aggregate-pomp-pompList.R
\name{fortify_aggregate.pompList}
\alias{fortify_aggregate.pompList}
\title{Takes in data from the R pomp package  where the output is a data frame and
puts it in SIR format for EpiCompare}
\usage{
\method{fortify_aggregate}{pompList}(
  data,
  states = NULL,
  enquo_states = NULL,
  package_source = NULL
)
}
\arguments{
\item{data}{Output from a pomp simulation where the output is a data frame,
\code{pomp::simulate()}}

\item{states}{vector of state names}

\item{enquo_states}{vector of quoted names}

\item{package_source}{optional package name}
}
\value{
data frame with the following columns
\describe{
\item{t}{the time}
\item{sim}{simulation number (factor variable) (optional column)}
\item{Xk}{where k = 0, ..., K}
}
}
\description{
Takes in data from the R pomp package  where the output is a data frame and
puts it in SIR format for EpiCompare
}
\details{
The default variables that are retained are SIR, but can be modified
  with the \code{states} argument.  If code{states = NULL}, we will attempt
  to find all single letter names in POMP and output those.
}
\examples{
out <- fortify_aggregate(pomp_arr, package_source = "pomp",
states = c("S", "I", "R"))
head(out)
unique(rowSums(out[, 3:5]))
}
