% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify-aggregate-pomp-list.R
\name{fortify_aggregate_ext.pomp_list}
\alias{fortify_aggregate_ext.pomp_list}
\title{Takes in data from the R pomp package  where the output is a data frame and puts it in SIR format for timeternR}
\usage{
\method{fortify_aggregate_ext}{pomp_list}(
  ext_data,
  states = c("S", "I", "R"),
  package_source = NULL
)
}
\arguments{
\item{ext_data}{Output from a pomp simulation where the output is a data frame, \code{pomp::simulate()}}

\item{states}{vector of state names}

\item{package_source}{optional package name}
}
\value{
data frame with the following columns
\describe{
\item{t}{the time}
\item{sim}{simulation number (factor variable) (optional column)}
\item{Xk}{where k = 0, ..., K}
}
}
\description{
Takes in data from the R pomp package  where the output is a data frame and puts it in SIR format for timeternR
}
\details{
The default variables that are retained are SIR, but can be modified with the \code{states} argument.  If code{states = NULL}, we will attempt to find all single letter names in POMP and output those.
}
