% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify-epimodel.R, R/fortify_epimodel.R
\name{fortify_EpiModel.icm}
\alias{fortify_EpiModel.icm}
\title{Takes in output from the \code{R} \code{EpiModel} package in \code{icm} format and puts it in SIR format}
\usage{
\method{fortify_EpiModel}{icm}(data)

\method{fortify_EpiModel}{icm}(data)
}
\arguments{
\item{data}{output from  \code{EpiModel::icm}}
}
\value{
data frame with the following columns
\describe{
\item{t}{the time}
\item{S}{number of Susceptibles}
\item{I}{number of Infectious}
\item{R}{number of Recovered}
\item{sim}{simulation number (factor variable) (optional column)}
}

data frame with the following columns
\describe{
\item{t}{the time}
\item{S}{number of Susceptibles}
\item{I}{number of Infectious}
\item{R}{number of Recovered}
\item{sim}{simulation number (factor variable) (optional column)}
}
}
\description{
Takes in output from the \code{R} \code{EpiModel} package in \code{icm} format and puts it in SIR format

Takes in output from the \code{R} \code{EpiModel} package in \code{icm} format and puts it in SIR format
}
\details{
Take the output from \code{EpiModel::icm} and turn it into an SIR data.frame for plotting.

Take the output from \code{EpiModel::icm} and turn it into an SIR data.frame for plotting.
}
\examples{
## For icm
sir <- fortify_EpiModel(EpiModel_icm)
head(sir)
class(sir)
## For icm
sir <- fortify_EpiModel(EpiModel_icm)
head(sir)
class(sir)
}
