% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify-epimodel.R
\name{fortify_epimodel}
\alias{fortify_epimodel}
\title{Generic method that takes in data from the \code{EpiModel} package and puts it in aggregate, SIR format}
\usage{
fortify_epimodel(data)
}
\arguments{
\item{data}{Output from a EpiModel simulation of class \code{icm} or \code{dcm}}
}
\value{
data frame with the following columns
\describe{
\item{t}{the time}
\item{Xk}{where k = 0, ..., K.}
\item{sim}{simulation number (factor variable) (optional column)}
}
}
\description{
Generic method that takes in data from the \code{EpiModel} package and puts it in aggregate, SIR format
}
\examples{
## For icm
sir <- fortify_EpiModel(EpiModel_icm)
head(sir)
class(sir)
}
