% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify_data.R
\name{fortify_pomp.list}
\alias{fortify_pomp.list}
\title{Takes in data from the R pomp package  where the output is array and puts it in SIR format for timeternR}
\usage{
\method{fortify_pomp}{list}(pomp_output)
}
\arguments{
\item{pomp_output}{Output from a pomp simulation where the output is 'array', \code{pomp::simulate()}}
}
\value{
data frame with the following columns
\describe{
\item{t}{the time}
\item{S}{number of Susceptibles}
\item{I}{number of Infectious}
\item{R}{number of Recovered}
\item{sim}{simulation number (factor variable) (optional column)}
}
}
\description{
Takes in data from the R pomp package  where the output is array and puts it in SIR format for timeternR
}
\details{
We require that the variables "S", "I", and "R" must be states in the pomp output.  Moreover, we will assume that these are the only relevant variables in the SIR calculation.
}
