% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify_data.R
\name{fortify_pomp}
\alias{fortify_pomp}
\title{Takes in data from the R pomp package and puts it in SIR format}
\usage{
fortify_pomp(pomp_output)
}
\arguments{
\item{pomp_output}{Output from a pomp simulation, \code{pomp::simulate()}}
}
\value{
data frame with the following columns
}
\description{
Takes in data from the R pomp package and puts it in SIR format
}
\details{
\describe{
\item{t}{}
}
}
