% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify_data.R
\name{fortify_siminf}
\alias{fortify_siminf}
\title{Takes in data from the R SimInf package and puts it in SIR format for plotting}
\usage{
fortify_siminf(siminf_output)
}
\arguments{
\item{siminf_output}{Output from a pomp simulation}
}
\value{
data frame with the following columns
}
\description{
Takes in data from the R SimInf package and puts it in SIR format for plotting
}
