% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conformal_prediction.R
\name{functional_psuedo_density_mode_cluster2}
\alias{functional_psuedo_density_mode_cluster2}
\title{Interatively apply the mean-shift algorithm to filamental objects}
\usage{
functional_psuedo_density_mode_cluster2(
  X_list,
  G_list = X_list,
  sigma,
  eps = 1e-05,
  maxT = 10,
  verbose = TRUE,
  usefrac = FALSE
)
}
\arguments{
\item{X_list}{list of filamental objects (data frames or matrices) 
that define the psuedo-density estimation}

\item{G_list}{another list of filament objects (data frames or matrices) 
that we will "walk" up to their modes.}

\item{sigma}{scalar, scale value for the distances between observations}

\item{eps}{scalar, if difference between steps is less than this - treat as 
if the point has converged}

\item{maxT}{integer, max number of iterations of the algorithm}

\item{verbose}{boolean, if we should show progress (done with a  C++ 
based progressbar)}

\item{usefrac}{boolean, if the distance should be weighted by the number
of points in the filament.}
}
\value{
Final step of points in the \cite{G_list} (hopefully the associated
modes).
}
\description{
Wrapper of C++ function `psuedo_density_mode_cluster2``
}
