% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conformal_prediction.R
\name{functional_psuedo_density_mode_cluster_r}
\alias{functional_psuedo_density_mode_cluster_r}
\title{An R (no cpp) version of `functional_psuedo_density_mode_cluster`}
\usage{
functional_psuedo_density_mode_cluster_r(
  X_list,
  G_list = X_list,
  sigma,
  eps = 1e-06,
  maxT = 10,
  verbose = TRUE
)
}
\arguments{
\item{X_list}{data list}

\item{G_list}{points to move}

\item{sigma}{sigma value}

\item{eps}{error term}

\item{maxT}{max number of steps}

\item{verbose}{boolean if we should be verbose}
}
\value{
list of updated filaments for points in G_list
}
\description{
This is not exported and is only included to a R sleuth to see what the 
function (that is written in C++ looks like in R). This won't be updated
even if the C++ code is (as of April 7, 2021).
}
\details{
This function is also used to test out the C++ version of the function
}
