% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_files_names.R
\name{gather_files_names}
\alias{gather_files_names}
\title{Make file names}
\usage{
gather_files_names(paths, type, nfcore_cutandrun)
}
\arguments{
\item{paths}{Character vector of file paths.}

\item{type}{File type to search for. Options include:
\itemize{
\item{"<pattern>"}{Finds files matching an arbitrary regex pattern
specified by user.}
\item{"peaks.stringent"}{Finds files ending in "*.peaks.bed.stringent.bed$"}
\item{"peaks.consensus"}{Finds files ending in "*.peaks.bed.stringent.bed$"}
\item{"peaks.consensus.filtered"}{
Finds files ending in"*.consensus.peaks.filtered.awk.bed$"}
\item{"picard"}{Finds files ending in
"*.target.markdup.MarkDuplicates.metrics.txt$"}
}}

\item{nfcore_cutandrun}{Whether the files were generated by the
\href{https://nf-co.re/cutandrun}{nf-core/cutandrun} Nextflow pipeline.
 If \code{TRUE}, can use the standardised folder structure to
 automatically generate more descriptive file names with sample IDs.}
}
\value{
Character vector of file names.
}
\description{
Support function for \link[EpiCompare]{gather_files}.
}
\keyword{internal}
