% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chrHMM_annotation.R
\name{get_chrHMM_annotation}
\alias{get_chrHMM_annotation}
\title{Download ChromHMM annotation file}
\usage{
get_chrHMM_annotation(cell_line)
}
\arguments{
\item{cell_line}{ChromHMM annotation for user-specified cell-line}
}
\value{
cell-line specific ChromHMM annotation file. Default K562 cell-line. Cell-line options are:
\itemize{
  \item "K562" = K-562 cells
  \item "Gm12878" = Cellosaurus cell-line GM12878
  \item "H1hesc" = H1 Human Embryonic Stem Cell
  \item "Hepg2" = Hep G2 cell
  \item "Hmec" = Human Mammary Epithelial Cell
  \item "Hsmm" = Human Skeletal Muscle Myoblasts
  \item "Huvec" = Human Umbilical Vein Endothelial Cells
  \item "Nhek" = Normal Human Epidermal Keratinocytes
  \item "Nhlf" = Normal Human Lung Fibroblasts
}
}
\description{
Download ChromHMM annotation file
}
