% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bands.R
\name{get_closest_nn}
\alias{get_closest_nn}
\title{create a grid of points indicating near border or not (and inside or outside)
using \code{RANN::nn2}.}
\usage{
get_closest_nn(
  border_points,
  inner_points,
  delta,
  xrange = NULL,
  yrange = NULL,
  gridbreaks = 100
)
}
\arguments{
\item{border_points}{data.frame of points from delta ball approach that are
"border points"}

\item{inner_points}{data.frame of points from delta ball approach that are
interior points.}

\item{delta}{float, size of delta ball radius}

\item{xrange}{vector, ordered values to examine in the x dimension (default
is NULL - will then be created using gridbreaks)}

\item{yrange}{vector, ordered values to examine in the y dimension (default
is NULL - will then be created using gridbreaks)}

\item{gridbreaks}{int, number of gridpoint in x and y dimensions if xrange or
yrange is not provided}
}
\value{
data frame, with \code{expand.grid} of xrange, yrange in columns
  \code{x} and \code{y} and a column \code{z} that indicates if it is: (1)
  not within delta to border points or inner_points, (2) if closest to
  border_points and (3) if closest to an inner_point.
}
\description{
This is internal code using \code{RANN::nn2} (faster).
}
