% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conformal_prediction.R, R/delta-ball.R
\name{get_delta_nn}
\alias{get_delta_nn}
\title{calculate maxmin distance between points}
\usage{
get_delta_nn(data)

get_delta_nn(data)
}
\arguments{
\item{data}{data.frame (with only columns that are needed)}

\item{df}{data.frame (with only columns that are needed)}
}
\value{
minimum radius for all points to be covered

minimum radius for all points to be covered
}
\description{
Fast calculation of maxmin distance using kd trees and nearest neighbors from
the \code{RANN} package.

Fast calculation of maxmin distance using kd trees and nearest neighbors from
the \code{RANN} package.
}
\details{
Taken from \code{simulationBands} package (written by Ben)
}
