% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify-aggregate-epimodel-icm.R
\name{get_epimodel_icm_states}
\alias{get_epimodel_icm_states}
\title{Get the viable names for the states from an icm object}
\usage{
get_epimodel_icm_states(data)
}
\arguments{
\item{data}{of class (icm), output from the EpiModel package}
}
\value{
vector of names of format {x}.num where {x} is a SINGLE letter
}
\description{
Get the viable names for the states from an icm object
}
