% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bands.R
\name{get_grid_elipsoid_containment}
\alias{get_grid_elipsoid_containment}
\title{create a grid of points indicating whether in a set of elipsoids}
\usage{
get_grid_elipsoid_containment(inside_func_list, xrange, yrange)
}
\arguments{
\item{inside_func_list}{list of functions that assess if observation is
in elipsoid (technically just need T/F function here...)}

\item{xrange}{vector, ordered values to examine in the x dimension}

\item{yrange}{vector, ordered values to examine in the y dimension}
}
\value{
updated_gridpoints (defined by yrange, xrange) with indication column
\code{included} if gridpoint is contained.
}
\description{
create a grid of points indicating whether in a set of elipsoids
}
\details{
See \code{check_inside_elipsoid} for functional idea
}
