% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-agents.R
\name{get_previous_counts}
\alias{get_previous_counts}
\title{Count number in each state}
\usage{
get_previous_counts(agent_states, n_states)
}
\arguments{
\item{agent_states}{vector of length agent_states where each has a value of -1, 0, ..., K-1.}

\item{n_states}{number of total states since we will use zeros}
}
\value{
We return a vector of length K of counts, the number in state 0 through K-S
}
\description{
Count number in each state
}
