% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-agents.R
\name{get_state_names}
\alias{get_state_names}
\title{Get the state names fro the transition matrix}
\usage{
get_state_names(trans_mat)
}
\arguments{
\item{trans_mat}{KxK transition matrix}
}
\value{
names of states
}
\description{
Get the state names fro the transition matrix
}
