% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/containment-and-band-creation.R
\name{grab_top_depth_filaments}
\alias{grab_top_depth_filaments}
\title{Select a proportion of the top / most deep filaments}
\usage{
grab_top_depth_filaments(
  grouped_df,
  data_columns = NULL,
  filament_depth_function = filament_distance_depth,
  alpha_level = 0.95
)
}
\arguments{
\item{grouped_df}{grouped_df data.frame object (assumed rows per filament are
ordered) - grouped per each filament}

\item{data_columns}{columns (currently strings) of data.frame that relate to
the filament's coordinates in euclidean space.}

\item{filament_depth_function}{function to calculate depth relative to the
filaments in the \code{grouped_df}. Will take in \code{data_columns}
parameter as well}

\item{alpha_level}{proportion of filaments to keep}
}
\value{
updated \code{grouped_df} with only the top depth filaments
}
\description{
similar to \code{depth_curves_to_points.grouped_df}
}
\examples{
library(dplyr)
top_filaments <- timeternR::pomp_df \%>\% group_by(.id) \%>\%
  filter(.id <= 10) \%>\%
  grab_top_depth_filaments(data_columns =c("S","I","R"),
                           alpha_level = .5)

# note that the below doesn't contain all filaments - as some have 0 depth
all_but_extreme_filaments <- timeternR::pomp_df \%>\% group_by(.id) \%>\%
  filter(.id <= 10) \%>\%
  grab_top_depth_filaments(data_columns = c("S","I","R"),
                           alpha_level = 0)
}
