% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hagelloch_agents}
\alias{hagelloch_agents}
\title{Measles in Hagelloch, Germany, 1861 (agent format)}
\format{A data frame with 188 rows and 3 columns. Below are the description of
each column.
\describe{
  \item{init_state}{Initial state (\code{factor}) for individual (at time t = 0). For the
  states, 0 = S, 1 = I, 2 = R.}
  \item{max_time_S}{maximum time individual was suspectable (S)}
  \item{max_time_I}{maximum time individual was infected (I)}
}}
\usage{
hagelloch_agents
}
\description{
This data is the agent formulation of 188 cases of measles among children in
the German city of Hagelloch, 1861, and is a summarization of the data in
\code{\link{hagelloch_raw}}. Each column is asssociated with 1 agent and
is a "sufficient" statistic for each agent's infection.  Each agent's
infection is uniquely identified by an initial state, max time before
infection (or time T), and max time before recovery (or time T). Note time T
is the final recorded time in the SIR model (in this case T = 94).
}
\examples{
## show first few cases
head(hagelloch_agents)
}
\keyword{datasets}
