% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hagelloch_raw}
\alias{hagelloch_raw}
\title{Measles in Hagelloch, Germany, 1861}
\format{A data frame with 188 rows and 12 columns
\describe{
  \item{PN}{Patient number (Case ID) per individual}
  \item{NAME}{Patient's last name}
  \item{FN}{Family number (individuals with matching \code{FN} are in the
  same family)}
  \item{AGE}{Age in years (fractions ignored)}
  \item{SEX}{Gender of the individual (factor: female; male)}
  \item{PRO}{\code{Date} of onset of prodromal symptoms}
  \item{ERU}{\code{Date} of onset of rash (not 100 sure of this one)}
  \item{CL}{School class (factor: preschool; 1st class; 2nd class )}
  \item{DEAD}{.}
  \item{IFTO}{.}
  \item{SI}{.}
  \item{C}{Complications (factor: "no complications", or what complication)}
  \item{PR}{.}
  \item{CA}{.}
  \item{NI}{.}
  \item{GE}{.}
  \item{TD}{.}
  \item{TM}{.}
  \item{x.loc}{x coordinate of house (in metres). Scaling in metres
  is obtained by multiplying the original coordinates by 2.5 (see details
  in Neal and Roberts (2004))}
  \item{y.loc}{y coordinate of house (in metres). See \code{x.loc} above}
  \item{tPRO}{exact time of onset of prodromal symptoms (relative to start
  of outbreak)}
  \item{tERU}{exact time of onset of rash (relative to start
  of outbreak) (not 100 sure of this one)}
  \item{tDEAD}{time of death (relative to start of outbreak) (\code{NA}
  implies recovered)}
  \item{tR}{exact time when individual entered into recovery}
  \item{tI}{exact time when individual was infected}
}}
\source{
\url{https://rdrr.io/cran/outbreaks/man/measles_hagelloch_1861.html}
}
\usage{
hagelloch_raw
}
\description{
These data comprise of 188 cases of measles among children in the German
city of Hagelloch, 1861. The data were
originally collected by Dr. Albert Pfeilsticker (1863) and augmented and
re-analysed by Dr. Heike Oesterle (1992). This data comes from the
\code{\link[outbreaks]{outbreaks}}.
}
\examples{
## show first few cases
head(hagelloch_raw)
}
\keyword{datasets}
