% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hagelloch_sir}
\alias{hagelloch_sir}
\title{Measles in Hagelloch, Germany, 1861 (SIR format)}
\format{
A data frame with 95 rows and 4 columns
\describe{
  \item{t}{time since outbreak, \eqn{t = 0, \dots, T=94}}
  \item{S}{Number of individuals suspectable}
  \item{I}{Number of individuals infected}
  \item{R}{Number of individuals in recovery}
}
}
\usage{
hagelloch_sir
}
\description{
This data is the SIR formulation of 188 cases of measles among children in
the German city of Hagelloch, 1861, and is a reformulation of the data in
\code{\link{hagelloch_raw}}. Each row is asssociated with 1 day (where
\code{t} tells the number of days after the start of the outbreak), and
the rest of the columns report the number of individuals suspectable,
infected, and in recovery.
}
\details{
Note that \eqn{s_t + i_t + r_t = 188} for each row \eqn{t in 0, \dots, 94}.
}
\examples{
## show first few cases
head(hagelloch_sir)
assertthat::assert_that(all(apply(hagelloch_sir[,-1], 1, sum) == 188))
}
\keyword{datasets}
