% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hg38_blacklist}
\alias{hg38_blacklist}
\title{Human genome hg38 blacklisted regions}
\format{
An object of class \code{GRanges} of length 636.
}
\source{
The code to prepare the .Rda file is:

\code{
hg38_blacklist <- ChIPseeker::readPeakFile(file.path(path), as = "GRanges")
usethis::use_data(hg38_blacklist, overwrite = TRUE)
}
}
\usage{
data("hg38_blacklist")
}
\description{
Obtained from \url{https://github.com/Boyle-Lab/Blacklist/tree/master/lists}.
The ENCODE blacklist includes regions of the genome that have anomalous and/
or unstructured signals.independent of the cell-line or experiment. Removal
of ENCODE blacklist is recommended for quality measure.
}
\keyword{datasets}
