% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_narrowPeak.R
\name{import_narrowPeak}
\alias{import_narrowPeak}
\title{Import narrow peaks}
\source{
\href{https://support.bioconductor.org/p/80535/#80537}{
Solution from Bioconductor forum}
}
\usage{
import_narrowPeak(
  path,
  as_file = file.path("https://www.encodeproject.org/documents",
    "948203bb-8eb2-42a2-8b12-1c10f356c998", "@@download/attachment", "narrowPeak.as"),
  extraCols = c(signalValue = "numeric", pValue = "numeric", qValue = "numeric", peak =
    "integer")
)
}
\arguments{
\item{path}{Path to local  \emph{narrowPeak} file or remote URL.}

\item{as_file}{Metadata describing each of the column names in
narrowPeak files. Provided by \href{https://www.encodeproject.org/}{ENCODE}.}

\item{extraCols}{What vector type to assign each of the extra columns to.}
}
\value{
\link[GenomicRanges]{GRanges} object.
}
\description{
Import \emph{narrowPeak} files generated by
\href{https://github.com/macs3-project/MACS}{\code{MACS}}.
}
\details{
Example dataset from:
 \url{https://www.encodeproject.org/files/ENCFF044JNJ/}.
 See \link[EpiCompare]{encode_H3K27ac} for more details.
}
\examples{
# path <- file.path("https://www.encodeproject.org/files",
#                   "ENCFF044JNJ/@download/ENCFF044JNJ.bed.gz")

path <- EpiCompare::write_example_peaks(datasets="encode_H3K27ac")
peaks <- EpiCompare::import_narrowPeak(path=path)
}
