% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{imports_hidden_from}
\alias{imports_hidden_from}
\title{import hidden function / variable from other package}
\usage{
imports_hidden_from(pkg, name)
}
\arguments{
\item{pkg}{package to grab hidden item from}

\item{name}{hidden item (function / variable)}
}
\value{
hidden object or function
}
\description{
see ?`:::`
}
