% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-agents.R
\name{initialize_agent_array}
\alias{initialize_agent_array}
\title{Initialize array for agent simulations}
\usage{
initialize_agent_array(
  init_vals,
  max_T,
  n_sims,
  birth_dates = NULL,
  birth_states = NULL
)
}
\arguments{
\item{init_vals}{vector of length K corresponding to the initial values in each state.  This will be over-ridden by birth_dates and birth_states if those are not NULL.}

\item{max_T}{total number of discrete time steps}

\item{n_sims}{number of simulations to run}

\item{birth_dates}{vector of size N, the maximum number of agents where each entry is the 'birth' of the agent into the system}

\item{birth_states}{which state an agent begins when born into the system}
}
\value{
n_sims x n_agents x (max_T + 1) array where entry ijt is the state of agent j in simulation i at time t-1.  The states they can take are integers between -1 (not born yet) and K-1 where 0 to K-1 correspond to different states.
}
\description{
Initialize array for agent simulations
}
