% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bands.R
\name{inner_compute_group_paths_to_points}
\alias{inner_compute_group_paths_to_points}
\title{inner function to compute points of top paths}
\usage{
inner_compute_group_paths_to_points(
  data,
  scales,
  params,
  pb_type = NULL,
  grid_size = rep(100, 2),
  conf_level = 0.9,
  over_delta = 0.1,
  dist_params = list(dist_approach = "auto", num_steps = "auto", quantile_approach =
    "depth", quantile_approach_params = list())
)
}
\arguments{
\item{data}{data as passed into geom_prediction_band (ggplot2 style)}

\item{scales}{scales as passed into geom_prediction_band (ggplot2 style)}

\item{params}{params as passed into geom_prediction_band (ggplot2 style)}

\item{pb_type}{string, see ?geom_prediction_band}

\item{grid_size}{vector of length 2, see ?geom_prediction_band}

\item{conf_level}{scalar (0-1), see ?geom_prediction_band}

\item{over_delta}{scalar, see ?geom_prediction_band}

\item{dist_params}{list, see ?geom_prediction_band}
}
\value{
data frame of points in the top paths (in simplex 3d projection)
}
\description{
projects paths to the unit simplex
}
