% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conformal_prediction.R
\name{inner_containment_conformal_score_mode_radius}
\alias{inner_containment_conformal_score_mode_radius}
\title{Helper fucntion for simulation-based conformal score calculation based 
potentially based on mode clustering and changing radius values.}
\usage{
inner_containment_conformal_score_mode_radius(
  df_row_group,
  simulations_group_df,
  data_column_names,
  simulation_info_df,
  list_radius_info,
  list_grouping_id,
  verbose = F
)
}
\arguments{
\item{df_row_group}{data frame with filaments information (can be a 
grouped_df) to be assessed.}

\item{simulations_group_df}{grouped data frame with multiple simulated 
filaments information}

\item{data_column_names}{columns of \code{df_row_group} and 
\code{simulations_group_df} that correspond to the output space.}

\item{simulation_info_df}{a dataframe with information of each simulation's
psuedo-density estimate, mode clustering, ranking of psuedo-density (within 
cluster and overall). See \code{inner_expanding_info} for expected structure.}

\item{list_radius_info}{list of lists of radius information per each mode.}

\item{list_grouping_id}{list of vectors of indices (grouped by mode cluster)
and ordered by psuedo-density values. These values are relative to the row
indices in simulation_info_df.}

\item{verbose}{boolean, if progress should be tracked with a progress bar}
}
\value{
data.frame with a row for each filament in \code{df_row_group} with a 
section column \code{containment_val} which is the discrete simulation-based
conformal score.
}
\description{
Helper fucntion for simulation-based conformal score calculation based 
potentially based on mode clustering and changing radius values.
}
