% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conformal_prediction.R
\name{inner_convert_single_radius_to_structure}
\alias{inner_convert_single_radius_to_structure}
\title{creates an increasing radius list structure for a fixed radius value.}
\usage{
inner_convert_single_radius_to_structure(radius, group_structure)
}
\arguments{
\item{radius}{scalar, single radius value}

\item{group_structure}{list of lists with a vector that captures the number
of observations in the group.}
}
\value{
returns list of lists (per each group) with min_cover_vec (vector of
radius value) and the dist_mat (radius per each obs included per step (col))
}
\description{
creates an increasing radius list structure for a fixed radius value.
}
