% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conformal_prediction.R
\name{inner_cumMax}
\alias{inner_cumMax}
\title{cumulative max of vector}
\usage{
inner_cumMax(x)
}
\arguments{
\item{x}{numerical vector (not checked)}
}
\value{
cumulative maximum of vector
}
\description{
cumulative max of vector
}
