% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta-ball.R
\name{inner_delta_ball_wrapper}
\alias{inner_delta_ball_wrapper}
\title{Run delta ball analysis}
\usage{
inner_delta_ball_wrapper(
  data_raw,
  xy_columns = c("x", "y"),
  n_steps = 100,
  remove_duplicates = F
)
}
\arguments{
\item{data_raw}{data frame with center points of balls}

\item{xy_columns}{columns of data.frame that relate to
  the points's coordinates in euclidean space. The input should look like
something like \code{c(x,y)} or \code{c("x","y")}.}

\item{n_steps}{number of equidistance points along the line, past delta on
both sides, that will be checked to approximate all points along the line}

\item{remove_duplicates}{boolean if need to remove duplicates in data_raw}
}
\value{
data frame of exterior lines (not ordered)
}
\description{
Run delta-ball analysis to obtain an outline points for
  delta-ball covering where points are in 2d space. We first find the minimum
  delta such that all balls centered at each point in the data set is
  touching at least 1 other ball (see \code{\link{get_delta}} for more
  information). This function then creates a geometric objects that trying to
  represent the covering of all these delta balls. Specifically we use
  geometric properties to find the points on the "outside" of this covering
  and return a set of lines that create a "boundary" of shorts. Intuition
  from "Computing Polygonal Surfaces from Unions of Balls" by Tam and
  Heidrich was used in this function.
}
\details{
This function (renamed as \code{delta_ball_wrapper}) is shared with \pkg{TCpredictionbands} on github:
\href{https://github.com/Mr8ND/TC-prediction-bands/tree/master/TCpredictionbands}{TCpredictionbands}.
}
