% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conformal_prediction.R
\name{inner_expanding_info}
\alias{inner_expanding_info}
\title{Combines psuedo-density estimate data frames and mode groupings and gets 
ordering per mode as well as globally}
\usage{
inner_expanding_info(psuedo_density_df, mode_grouping)
}
\arguments{
\item{psuedo_density_df}{data frame with object information and 
psuedo-density estimates (in column called \code{psuedo_density})}

\item{mode_grouping}{data frame with object information and 
mode clustering index (in column called \code{groupings})}
}
\value{
a list of 2 objects. The first is a data frame defined by the joining
of the previous data frames and additional columns for \code{ranking} of 
their psuedo_densities and for \code{group_ranking} (ranking conditional on 
mode group). The second element of the list is a lists of indicies associated
with the \code{group_ranking} making to object interest.
}
\description{
Ranking ties are randomly assigned index values
}
