% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thesis_data.R
\name{interior_clustering}
\alias{interior_clustering}
\title{Interior function for mode clustering across a range of eps given a single
maxT value}
\usage{
interior_clustering(x_list_inner, g_list_inner, sigma, range_eps, maxT)
}
\arguments{
\item{x_list_inner}{list of curves to define psuedo-density}

\item{g_list_inner}{list of curves to walk up the psuedo-density}

\item{sigma}{scalar for the sigma of the psuedo-density}

\item{range_eps}{vector and eps value (descending in size)}

\item{maxT}{int, max number of iterations}
}
\value{
linked list for each eps explored before maxT = 0.
}
\description{
Interior function for mode clustering across a range of eps given a single
maxT value
}
