% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-dist-mat.R
\name{is.tidy_dist_mat}
\alias{is.tidy_dist_mat}
\title{checks if object is a tidy_dist_mat}
\usage{
is.tidy_dist_mat(x)
}
\arguments{
\item{x}{any old object}
}
\value{
boolean logic if object is a \code{tidy_dist_mat} object
}
\description{
checks if object is a tidy_dist_mat
}
