% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_granges.R
\name{is_granges}
\alias{is_granges}
\title{Is an object of class GRanges}
\usage{
is_granges(obj)
}
\arguments{
\item{obj}{Any R object.}
}
\value{
Boolean.
}
\description{
Check whether an object is of the class \link[GenomicRanges]{GRanges}.
}
\keyword{internal}
