% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde.R
\name{kde_from_list}
\alias{kde_from_list}
\title{Contour of data points using Kernel Density Estimate.}
\usage{
kde_from_list(
  dflist,
  alpha,
  h_band = NULL,
  position = 1:2,
  grid_size = rep(1000, 2)
)
}
\arguments{
\item{dflist}{list of data frames}

\item{alpha}{contour level, an scalar from the sequence .01 to .99 by step
size .01. For example, a value of .95 gives a contour level associated with
the cumulative prob of 95\% (and 95\% confidence interval)}

\item{h_band}{optional argument for the bandwidth of the kde object. If NULL,
the optimal band would be selected through the \code{\link[ks:kde]{ks::kde}}
function. Default is NULL.}

\item{position}{Columns position of x/y pair. Default is 1:2}

\item{grid_size}{size of the grid which is going to be used for the
evaluation of kde object. Can be reduced to speed-up computation.}
}
\value{
\code{contour} List of contour(s) at the specified level
}
\description{
This function calculates contour points and area from list of data.frames
with data in 2d. Basically this is a wrapper of \code{\link[ks:kde]{ks::kde}}.
}
\details{
This function (renamed as \code{kde_from_tclist}) is shared with
\pkg{TCpredictionbands} on github:
\href{https://github.com/Mr8ND/TC-prediction-bands/tree/master/TCpredictionbands}{TCpredictionbands}.
}
