% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conformal_prediction.R
\name{l2_filament_distance}
\alias{l2_filament_distance}
\title{inner l2 distance between filaments}
\usage{
l2_filament_distance(df1, df2, usefrac = F)
}
\arguments{
\item{df1}{data.frame (n x p)}

\item{df2}{data.frame (n x p)}

\item{usefrac}{if we should calculate the distance relative to a scaling of 
1/nrow(\code{df1}) (before taking the square-root).}
}
\value{
distance between filaments
}
\description{
Basically the same as \code{l2filamentdist_df}
}
