% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance-between.R
\name{l2filamentdist_df}
\alias{l2filamentdist_df}
\title{inner l2 distance between filaments}
\usage{
l2filamentdist_df(m1, m2)
}
\arguments{
\item{m1}{data.frame (n x p)}

\item{m2}{data.frame (n x p)}
}
\value{
distance between filaments
}
\description{
inner l2 distance between filaments
}
