% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liftover_grlist.R
\name{liftover_grlist}
\alias{liftover_grlist}
\title{Liftover peak list}
\usage{
liftover_grlist(
  grlist,
  input_build,
  output_build = "hg19",
  as_grangeslist = FALSE
)
}
\arguments{
\item{grlist}{A named list of \link[GenomicRanges]{GRanges} object.}

\item{input_build}{The genome build of \code{grlist}.}

\item{output_build}{Desired genome build for
\code{grlist} to be lifted over to.}

\item{as_grangeslist}{Convert output to class
\link[GenomicRanges]{GRangesList} before returning.}
}
\value{
Named list of lifted \link[GenomicRanges]{GRanges} objects.
}
\description{
Perform genome build liftover to on a named list of
 \link[GenomicRanges]{GRanges} objects.
}
\examples{
data("encode_H3K27ac") # example dataset as GRanges object
data("CnT_H3K27ac") # example dataset as GRanges object
data("CnR_H3K27ac") # example dataset as GRanges object
grlist <- list("encode_H3K27ac"=encode_H3K27ac,
               "CnT_H3K27ac"=CnT_H3K27ac,
               "CnR_H3K27ac"=CnR_H3K27ac)

grlist_lifted <- liftover_grlist(grlist = grlist,
                                 input_build = "hg19",
                                 output_build="hg38")
}
