% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mm9_blacklist}
\alias{mm9_blacklist}
\title{Mouse genome mm9 blacklisted regions}
\format{
An object of class \code{GRanges} of length 292.
}
\source{
\code{
tmp <- base::get("mm10_blacklist", asNamespace("EpiCompare"))
mm9_blacklist <- liftover_grlist(grlist = tmp,
                                 input_build = "mm10",
                                 output_build = "mm9", 
                                 keep_chr = NULL)
usethis::use_data(mm9_blacklist, overwrite = TRUE)                                
}
}
\usage{
data("mm9_blacklist")
}
\description{
Blaklisted regions of the mm9 genome build 
btained by lifting over the \code{mm10_blacklist}.
}
\keyword{datasets}
