% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thesis_data.R
\name{mode_cluster2}
\alias{mode_cluster2}
\title{smart model clustering that examines mode clustering across a range of 
paramters}
\usage{
mode_cluster2(
  g_list,
  g_names,
  position,
  sigma,
  range_eps = 10^seq(-5, -8, by = 0.5),
  range_maxT = c(seq(30, 130, by = 20), 200),
  range_diff_eps = 10^seq(-5, -8, by = 0.5),
  usefrac = T,
  verbose = T,
  collect_group_info = F
)
}
\arguments{
\item{g_list}{list of initial objects}

\item{g_names}{data frame of names of the initial objects}

\item{position}{positions in g_list that define the trajectory}

\item{sigma}{scalar for the sigma of the psuedo-density}

\item{range_eps}{vector of eps value (descending in size) - to be used in the
mode accession}

\item{range_maxT}{vector of maxT values (increasing in size)}

\item{range_diff_eps}{vector of diff eps value (descending in size) - to 
be using in calculating the number of groups}

\item{usefrac}{boolean, if distance function should multiple by number of 
points used}

\item{verbose}{verbose boolean, if should report progress}
}
\value{
data frame with parameter values and number of clusters found
}
\description{
smart model clustering that examines mode clustering across a range of 
paramters
}
