% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conformal_prediction.R
\name{mode_clustering}
\alias{mode_clustering}
\title{Find mode clusters of filament objects}
\usage{
mode_clustering(
  g_list,
  g_names,
  position,
  sigma,
  eps = 1e-05,
  maxT = 30,
  diff_eps = 1e-05,
  usefrac = F,
  verbose = T
)
}
\arguments{
\item{g_list}{list of filamental objects (data frames or matrices) 
that we'd like to cluster (will be clustered based on the psuedo-density
estimate relative to themselves.)}

\item{g_names}{grouping information (as a data frame) asssociated with 
elements on the \code{g_list}}

\item{position}{vector of indices of columns that will be used to define the 
location of the points in space}

\item{sigma}{scalar, scale value for the distances between observations}

\item{eps}{scalar, if difference between steps is less than this - treat as 
if the point has converged}

\item{maxT}{integer, max number of iterations of the algorithm}

\item{diff_eps}{scalar, if the final step of each of the points is within
this distance from each-other they will be grouped together.}

\item{usefrac}{boolean, if we should use fraction approach to l2 distance
or not. Naturally is related to \code{sigma} value.}

\item{verbose}{boolean, if we should show progress (done with a  C++ 
based progressbar)}
}
\value{
data frame with \code{g_names} and a \code{grouping} column which 
indicates which mode group each observation is in.
}
\description{
Wrapper of C++ function `psuedo_density_mode_cluster`` plus some cluster
finding
}
