% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-agents-multinomial-updater.R
\name{multinomial_updater}
\alias{multinomial_updater}
\title{Multinomial draws to update states depending on past count}
\usage{
multinomial_updater(state_counts, par_vals, D_list_fxns, time = NA)
}
\arguments{
\item{state_counts}{vector of length K of counts in each state}

\item{par_vals}{parameter vector values}

\item{D_list_fxns}{output from \code{\link{trans_mat_to_D_fxn}}}

\item{time}{current time, possibly used in the udpate function}
}
\value{
KxK matrix with new number in each state for agents who were in state i
}
\description{
Multinomial draws to update states depending on past count
}
