% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-dist-mat.R
\name{not}
\alias{not}
\alias{not.data.frame}
\alias{is.not_df}
\title{characterizes that we will subset by the indices *not* in said data.frame}
\usage{
not(x)

\method{not}{data.frame}(x)

is.not_df(x)
}
\arguments{
\item{x}{data.frame like object with index structure associated with a 
\code{tidy_dist_mat}'s rownames or colnames.}
}
\value{
a new \code{not_df} object
}
\description{
characterizes that we will subset by the indices *not* in said data.frame
}
