% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap_percent.R
\name{overlap_percent}
\alias{overlap_percent}
\title{Calculate percentage of overlapping peaks}
\usage{
overlap_percent(peaklist, reference, invert = FALSE)
}
\arguments{
\item{peaklist}{A list of peak files as GRanges object.
Files must be listed using `list()` and named using `names()`
If not named, default file names will be assigned.}

\item{reference}{refrence peaks file}

\item{invert}{whether to invert}
}
\value{
data frame
}
\description{
Calculate percentage of overlapping peaks
}
