% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap_stat_plot.R
\name{overlap_stat_plot}
\alias{overlap_stat_plot}
\title{Statistical significance of overlapping peaks}
\usage{
overlap_stat_plot(reference, peaklist, annotation = NULL)
}
\arguments{
\item{reference}{A reference peak file as GRanges object.}

\item{peaklist}{A list of peak files as GRanges object.
Files must be listed and named using \code{list()}.
E.g. \code{list("name1"=file1, "name2"=file2)}.
If not named, default file names will be assigned.}

\item{annotation}{A TxDb annotation object from Bioconductor. This is
required only if the reference file does not have BED6+4 format.}
}
\value{
A boxplot or barplot showing the statistical significance of
overlapping/non-overlapping peaks.
}
\description{
This function calculates the statistical significance of overlapping/
non-overlapping peaks against a reference peak file.If the reference peak
file has the BED6+4 format (peak called by MACS2), the function generates a
series of boxplots showing the distribution of q-values for sample peaks that
are overlapping and non-overlapping with the reference. If the reference peak
file does not have the BED6+4 format, the function uses `enrichPeakOverlap()`
from `ChIPseeker` package to calculate the statistical significance of
overlapping peaks only. In this case, please provide an annotation file as
TxDb object.
}
\examples{
### Load Data ###
data("encode_H3K27ac") # example peakfile GRanges object
data("CnT_H3K27ac") # example peakfile GRanges object
data("CnR_H3K27ac") # example peakfile GRanges object

### Create Named Peaklist & Reference ###
peaklist <- list('CnT'=CnT_H3K27ac, "CnR"=CnR_H3K27ac)
reference <- list("ENCODE"=encode_H3K27ac)

### Run ###
out <- overlap_stat_plot(reference = reference,
                         peaklist = peaklist)
stat_plot <- out[[1]] # plot
stat_df <- out[[2]] # df

}
