% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_info.R
\name{peak_info}
\alias{peak_info}
\title{Summary of Peak Information}
\usage{
peak_info(peak_list, blacklist)
}
\arguments{
\item{peak_list}{A named list of peak files as GRanges object.
Objects listed using `list()` and named using `names()`.}

\item{blacklist}{A GRanges object containing blacklisted regions.}
}
\value{
A summary table of peak information
}
\description{
This function outputs a table summarizing information on the peak files.
Provides the total number of peaks and the percentage of peaks in blacklisted regions.
}
\examples{
library(EpiCompare)
data("encode_H3K27ac") # example dataset as GRanges object
data("CnT_H3K27ac") # example dataset as GRanges object
data("hg19_blacklist") # example blacklist dataset as GRanges object

peaklist <- list(encode_H3K27ac, CnT_H3K27ac) # list two peakfiles
names(peaklist) <- c("encode", "CnT") # set names

peak_info(peak_list = peaklist,
          blacklist = hg19_blacklist)

}
