% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_chrHMM.R
\name{plot_chrHMM}
\alias{plot_chrHMM}
\title{Create ChromHMM Heatmap}
\usage{
plot_chrHMM(
  chrHMMfile1,
  chrHMMfile1_name,
  chrHMMfile2,
  chrHMMfile2_name,
  chrHMM_list
)
}
\arguments{
\item{chrHMMfile1}{peakfile1 as GRanges}

\item{chrHMMfile1_name}{name of peakfile1}

\item{chrHMMfile2}{peakfile2 as GRanges}

\item{chrHMMfile2_name}{name of peakfile2}

\item{chrHMM_list}{ChromHMM annotation list}
}
\value{
ChromHMM heatmap
}
\description{
Creates a heatmap using outputs from ChromHMM using ggplot2.
The function takes two peak files, performs ChromHMM and outputs a heatmap.
ChromHMM annotation file must be loaded prior to using this function.
}
\keyword{internal}
