% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_chrmHMM.R
\name{plot_chrmHMM}
\alias{plot_chrmHMM}
\title{Plot ChromHMM heatmap}
\usage{
plot_chrmHMM(peaklist, chrmHMM_annotation)
}
\arguments{
\item{peaklist}{A list of peak files as GRanges object.
Files must be listed using `list()` and named using `names()`
If not named, default file names will be assigned.}

\item{chrmHMM_annotation}{ChromHMM annotation list}
}
\value{
ChromHMM heatmap
}
\description{
Creates a heatmap using outputs from ChromHMM using ggplot2.
The function takes a list of peakfiles, performs ChromHMM and outputs a heatmap.
ChromHMM annotation file must be loaded prior to using this function.
}
