% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_enrichment.R
\name{plot_enrichment}
\alias{plot_enrichment}
\title{Generate enrichment analysis plots}
\usage{
plot_enrichment(peaklist, namelist)
}
\arguments{
\item{peaklist}{A list of peak files as GRanges object. Files must be listed using `list()`}

\item{namelist}{A list of file names in the same order as the list of peak files. Use `c()` for multiple.}
}
\value{
KEGG and GO dot plots
}
\description{
This function runs KEGG and GO enrichment analysis of peak files
and generates dot plots.
}
