% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pomp_arr}
\alias{pomp_arr}
\title{Example SIR simulation output of class 'array' from the \code{pomp} package}
\format{A list of 2 entries where the first and second entries are both arrays.  The first entry has dimension 4x 100 x10 and the dimension names are respectively variable ("S", "I", "R", "H"), rep, and time.  The second entry has dimension 1 x 100 x 101 where the dimension names are respectively variable ("cases"), 'rep', and 'time'.  Thus entry ijk in either array corresponds to the number of the variable i at simulation/rep j and time k.}
\usage{
pomp_arr
}
\description{
The data is an example of simulation output from functions from the \href{https://kingaa.github.io/pomp/}{pomp package by King et al}.  The code for how this data was generated can be found in the \code{data-raw} folder.  We also make available the other possible output formats from pomp.  See \link{pomp_df} and \link{pomp_pomp}
}
\examples{
data("pomp_arr")
names(pomp_arr)
dimnames(pomp_arr[[1]])
}
\keyword{datasets}
