% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pomp_df}
\alias{pomp_df}
\title{Example SIR simulation output of class ' data.frame' from the \code{pomp} package}
\format{
A data frame of dimension 10100 x 7 where the columns are
\describe{
\item{time}{an integer value between 0 and 100}
\item{.id}{the simulation ID number}
\item{S}{The number of Susceptible at given time and simulation ID}
\item{I}{The number of Infectious at given time and simulation ID}
\item{R}{The number of Recovered at given time and simulation ID}
\item{H}{A helper variable}
\item{cases}{}
}
}
\usage{
pomp_df
}
\description{
The data is an example of simulation output from functions from the \href{https://kingaa.github.io/pomp/}{pomp package by King et al}.  The code for how this data was generated can be found in the \code{data-raw} folder.  We also make available the other possible output formats from pomp.  See \link{pomp_arr} and \link{pomp_pomp}
}
\examples{
data("pomp_df")
head(pomp_df)
library(ggplot2)
ggplot(pomp_df) + geom_line(aes(x = time, y = I, group = .id,
col = as.numeric(.id)))
}
\keyword{datasets}
