% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pomp_pomp}
\alias{pomp_pomp}
\title{Example SIR simulation output of class ' pomp' from the \code{pomp} package}
\format{
The object is a 'pompList' which contains essentially all information from a \code{pomp} simulation.
}
\usage{
pomp_pomp
}
\description{
The data is an example of simulation output from functions from the \href{https://kingaa.github.io/pomp/}{pomp package by King et al}.  The code for how this data was generated can be found in the \code{data-raw} folder.  We also make available the other possible output formats from pomp.  See \link{pomp_df} and \link{pomp_arr}
}
\examples{
data("pomp_pomp")
str(pomp_pomp)
}
\keyword{datasets}
