% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precision_recall_matrix.R
\name{precision_recall_matrix}
\alias{precision_recall_matrix}
\title{Create a precision-recall matrix}
\usage{
precision_recall_matrix(peaklist, fill_diag = NA, verbose = TRUE)
}
\arguments{
\item{fill_diag}{Fill the diagonal of the overlap matrix.}

\item{verbose}{Print messages.}
}
\value{
matrix
}
\description{
Converts a list of peak files to a symmetric matrix where 
the y-axis indicates precision and the x-axis indicates recall.
}
\keyword{internal}
