% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_values.R
\name{predict_values}
\alias{predict_values}
\title{Predict values}
\usage{
predict_values(df, fun, values, input_var, predicted_var)
}
\arguments{
\item{df}{data.frame}

\item{fun}{Function to fit the data with.}

\item{values}{Values to make predictions from.}

\item{input_var}{Input variable column name.}

\item{predicted_var}{Predicted variable name.}
}
\value{
data.frame
}
\description{
Fit a model and make predictions from it.
}
\keyword{internal}
