% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_blacklist.R
\name{prepare_blacklist}
\alias{prepare_blacklist}
\title{Prepare blacklist as GRanges}
\usage{
prepare_blacklist(
  blacklist,
  output_build,
  blacklist_build = NULL,
  verbose = TRUE
)
}
\arguments{
\item{output_build}{Desired genome build for
\code{grlist} to be lifted over to.}

\item{blacklist_build}{Genome build of the blacklist.
Only used when \code{blacklist} is a user-supplied 
\link[GenomicRanges]{GRanges} object.}

\item{verbose}{Print messages.}
}
\value{
A \link[GenomicRanges]{GRanges} objects of blacklisted 
genomic regions from the relevant genome build.
}
\description{
Selects the appropriate blacklist in a variety of conditions.
}
\keyword{internal}
