% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_peaklist.R
\name{prepare_peaklist}
\alias{prepare_peaklist}
\title{Prepare peaklist as GRanges}
\usage{
prepare_peaklist(peaklist, remove_empty = TRUE, as_grangeslist = FALSE)
}
\arguments{
\item{peaklist}{A named list of peaks as \link[GenomicRanges]{GRanges}
or paths to BED files.}

\item{remove_empty}{Remove any empty elements in the list.}

\item{as_grangeslist}{Convert output to class
\link[GenomicRanges]{GRangesList} before returning.}
}
\value{
A list of \link[GenomicRanges]{GRanges} objects
}
\description{
Prepare peaklist as GRanges
}
\keyword{internal}
