% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_reference.R
\name{prepare_reference}
\alias{prepare_reference}
\title{Prepare referemce as GRanges}
\usage{
prepare_reference(reference, max_elements = NULL)
}
\arguments{
\item{reference}{A named list of \link[GenomicRanges]{GRanges} objects, 
or a single \link[GenomicRanges]{GRanges} object to be converted into a 
named list.}

\item{max_elements}{Max number of elements to use within the list.
Set to \code{NULL} (default) to use all elements.}
}
\value{
A list of \link[GenomicRanges]{GRanges} objects
}
\description{
Prepare referemce as GRanges
}
\keyword{internal}
