% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{print.ggplot}
\alias{print.ggplot}
\title{Explicitly draw plot (\code{ggtern} and \code{ggplot2} compatible)}
\usage{
\method{print}{ggplot}(x, newpage = is.null(vp), vp = NULL, ...)
}
\arguments{
\item{x}{plot to display}

\item{newpage}{draw new (empty) page first?}

\item{vp}{viewport to draw plot in}

\item{...}{other arguments not used by this method}
}
\value{
Invisibly returns the result of \code{ggplot_build()}, which is a 
list with components that contain the plot itself, the data, information 
about the scales, panels etc.
}
\description{
Makes sure both \code{\link[ggtern]{ggtern}} and 
\code{\link[ggplot2]{ggplot2}} objects 
display correctly with the \code{print.ggplot} function.
}
