% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-dist-mat.R
\name{print.tidy_dist_mat}
\alias{print.tidy_dist_mat}
\title{print tidy_dist_mat objects}
\usage{
\method{print}{tidy_dist_mat}(x, ..., n = NULL)
}
\arguments{
\item{x}{tidy_dist_mat object}

\item{...}{(like \code{digits = 6}) number of significant digits to display
(uses \code{signif})}
}
\description{
print tidy_dist_mat objects
}
\examples{
inner_data <- data.frame(x = rnorm(3), y = rnorm(3))
my_dist_mat <- as.matrix(dist(inner_data))
rownames_df <- data.frame(id = 1:3)
colnames_df <- data.frame(id = c(1,2,1), id2 = c("f", "f", "s"))

my_tidy_dm <- tidy_dist_mat(my_dist_mat, rownames_df, colnames_df)
print(my_tidy_dm)
}
