% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-dist-mat.R
\name{process_df_index}
\alias{process_df_index}
\title{internal function to get the indices of the \code{tidy_dist_mat} related to a data frame.}
\usage{
process_df_index(x, index, margin = 1)
}
\arguments{
\item{x}{\code{tidy_dist_mat} object}

\item{index}{data.frame style index related to \code{x}}

\item{margin}{scalar representing if \code{index} is related to rows or 
columns of \code{x}. 1 = row, 2 = column.}
}
\value{
index vector
}
\description{
internal function to get the indices of the \code{tidy_dist_mat} related to a data frame.
}
