% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bands.R
\name{project_to_simplex}
\alias{project_to_simplex}
\title{project onto a standard 3d simplex.}
\usage{
project_to_simplex(df_3d, column_names = c("x", "y", "z"))
}
\arguments{
\item{df_3d}{data frame}

\item{column_names}{names of columns for the 3 dimensions.}
}
\value{
an updated version of \code{df_3d} with points projected onto
simplex.
}
\description{
project onto a standard 3d simplex.
}
