% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{psuedo_density_mode_cluster}
\alias{psuedo_density_mode_cluster}
\title{psuedo density walking towards the mode}
\usage{
psuedo_density_mode_cluster(
  X_list,
  G_list,
  sigma,
  eps = 1e-06,
  maxT = 10L,
  verbose = TRUE,
  usefrac = FALSE
)
}
\arguments{
\item{X_list}{list of matrices (same size), these define the psuedo density}

\item{G_list}{list of matrices to walk up (can be the same as X_list, but 
need not be).}

\item{sigma}{double, sigma for distance psuedo density calculation}

\item{eps}{double, distance between iterative steps when to stop progressing 
(aka stop at a mode)}

\item{maxT}{int, max number of interations}

\item{verbose}{boolean, if we should use a progress bar}

\item{usefrac}{boolean, if we should calculate the distance relative to this 
scaling (should also impact the earlier calculation of sigma)}
}
\value{
Updated step of G_list
}
\description{
psuedo density walking towards the mode
}
