% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bowtie.R
\name{read_bowtie}
\alias{read_bowtie}
\title{Read bowtie}
\usage{
read_bowtie(path, verbose = TRUE)
}
\arguments{
\item{path}{Path to bowtie file.}

\item{verbose}{Print messages.}
}
\value{
\link[data.table]{data.table}
}
\description{
Read a bowtie file.
}
\keyword{internal}
