% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta-ball.R
\name{remove_incomplete_tri}
\alias{remove_incomplete_tri}
\title{Remove triangles with edges that need removal}
\usage{
remove_incomplete_tri(tuples_of_tri, removed_mat)
}
\arguments{
\item{tuples_of_tri}{data frame with tuples of triangle edges and
triangle index}

\item{removed_mat}{edges to be removed}
}
\value{
data frame with tuples of triangle not removed
}
\description{
Remove triangles from tuple matrix that have 1 or more edge that
needs to be removed
}
\details{
This function is shared with \pkg{TCpredictionbands} on github:
\href{https://github.com/Mr8ND/TC-prediction-bands/tree/master/TCpredictionbands}{TCpredictionbands}.
}
