% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_nonstandard_chrom.R
\name{remove_nonstandard_chrom}
\alias{remove_nonstandard_chrom}
\title{Remove non-standard chromosomes}
\usage{
remove_nonstandard_chrom(
  grlist,
  keep_chr = paste0("chr", c(seq_len(22), "X", "Y")),
  verbose = TRUE
)
}
\arguments{
\item{grlist}{Named list of \link[GenomicRanges]{GRanges} objects.}

\item{keep_chr}{Which chromosomes to keep.}

\item{verbose}{Print messages.}
}
\value{
Named list of \link[GenomicRanges]{GRanges} objects.
}
\description{
Remove non-standard chromosomes from a list of 
\link[GenomicRanges]{GRanges} objects.
}
\keyword{internal}
