% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_header.R
\name{report_header}
\alias{report_header}
\title{Report header}
\usage{
report_header()
}
\value{
Header string to be rendering within Rmarkdown file.
}
\description{
Generate a header for \link[EpiCompare]{EpiCompare} reports generated using 
the \emph{EpiCompare.Rmd} template.
}
\examples{
report_header()
}
