% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-dist-mat.R
\name{rownames<-.tidy_dist_mat}
\alias{rownames<-.tidy_dist_mat}
\title{assign rownames for tidy_dist_mat}
\usage{
\method{rownames}{tidy_dist_mat}(x) <- value
}
\arguments{
\item{x}{tidy_dist_mat}

\item{value}{updated rownames}
}
\description{
assign rownames for tidy_dist_mat
}
