% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_output.R
\name{save_output}
\alias{save_output}
\title{Save output}
\usage{
save_output(
  save_output = FALSE,
  file,
  file_type,
  filename,
  outpath,
  interactive = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{save_output}{Default FALSE. If TRUE, outputs are saved.}

\item{file}{Tables and plots to be saved.}

\item{file_type}{Type of file to be saved.
"data.frame", "ggplot", "image"}

\item{filename}{Name of file.}

\item{outpath}{Outpath}

\item{interactive}{Default FALSE. If TRUE, interactive plots are saved as
html.}

\item{verbose}{Print messages.}
}
\value{
Saved data frames and plots.
}
\description{
This function saves data frames and plots generated by EpiCompare.
}
\keyword{internal}
