% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{scalartimesmat}
\alias{scalartimesmat}
\title{multiples a matrix by a scalar}
\usage{
scalartimesmat(m, s)
}
\arguments{
\item{m}{matrix}

\item{s}{scalar (double)}
}
\value{
matrix who's entries are m's time s.
}
\description{
multiples a matrix by a scalar
}
