% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{SEIR_to_SIR_E}
\alias{SEIR_to_SIR_E}
\title{Convert SEIR to XYZ coordinates fixed in a tetrahedron}
\usage{
SEIR_to_SIR_E(data, ternary_vars = c("S", "I", "R"), group_var = "E")
}
\arguments{
\item{data}{data frame with the following columns
\describe{
\item{t}{time step}
\item{S}{Number of people in S}
\item{E}{Number of people in E}
\item{I}{Number of people in I}
\item{R}{Number of people in R}
}}

\item{ternary_vars}{named vector of the three variables to use as the sides of the ternary plot}

\item{group_var}{name of the variable to use as the color/feature/grouping vector}
}
\value{
data frame with the transformed variables SEIR -> s, i, r, group variables
}
\description{
Convert SEIR to XYZ coordinates fixed in a tetrahedron
}
\examples{
seir <- data.frame(t = 0:3,
S = c(90, 80, 70, 60),
E = c(0, 10, 10, 10),
I = c(10, 10, 10, 10),
R = c(0, 0, 10, 20))
seir_xyz <- SEIR_to_XYZ(seir)
head(seir_xyz)
}
