% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{SEIR_to_XYZ}
\alias{SEIR_to_XYZ}
\title{Convert SEIR to XYZ coordinates fixed in a tetrahedron}
\usage{
SEIR_to_XYZ(data, var_order = c("S", "E", "I", "R"))
}
\arguments{
\item{data}{data frame with the following columns
\describe{
\item{time}{time step}
\item{S}{Number of people in S}
\item{E}{Number of people in E}
\item{I}{Number of people in I}
\item{R}{Number of people in R}
}}

\item{var_order}{vector of column names corresponding to the different axes of the tetrahedron-based coordinate system:  (t, l, r, f) which stands for top, left, right, front.}
}
\value{
original data frame along with columns x, y, and z
}
\description{
Convert SEIR to XYZ coordinates fixed in a tetrahedron
}
\examples{
seir <- data.frame(t = 0:3,
S = c(90, 80, 70, 60),
E = c(0, 10, 10, 10),
I = c(10, 10, 10, 10),
R = c(0, 0, 10, 20))
seir_xyz <- SEIR_to_XYZ(seir)
head(seir_xyz)
}
