% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_min_max.R
\name{set_min_max}
\alias{set_min_max}
\title{Set min/max}
\usage{
set_min_max(df, colname, min_val = 0, max_val = 100)
}
\arguments{
\item{df}{data.frame}

\item{colname}{Column name to check.}

\item{min_val}{Minimum value.}

\item{max_val}{Maximum value.}
}
\value{
data.frame
}
\description{
Set the min/max values in a data.frame.
}
\keyword{internal}
