% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir-array-to-df.R
\name{sim_arr_to_df}
\alias{sim_arr_to_df}
\title{Convert simulation output from simulate_SIR_agents to data frame format}
\usage{
sim_arr_to_df(arr)
}
\arguments{
\item{arr}{array of dimension \eqn{\# agents \times \# sims \times 3} where agent ijk is the ith agent in simulation j, and looking at the kth column where 1 is the initial state, 1 is the time of infection, and 2 is the time of recovery}
}
\value{
data.frame of dimension (n_agents x n_sims) x 4 where the columns are
\describe{
\item{agent_id}{unique ID}
\item{sim}{simulation number}
\item{tI}{time of infection}
\item{tR}{time of recovery}
}
}
\description{
Convert simulation output from simulate_SIR_agents to data frame format
}
