% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conformal_prediction.R
\name{simulation_based_conformal3.5}
\alias{simulation_based_conformal3.5}
\title{global wrapper for simulation-based conformal score calculation that allows 
for mode clustering and changes of radius.}
\usage{
simulation_based_conformal3.5(
  truth_grouped_df,
  simulations_grouped_df,
  data_column_names = c("S", "I", "R"),
  number_points = Inf,
  .change_radius = TRUE,
  .mode_cluster = TRUE,
  .to_simplex = TRUE,
  .use_frac = FALSE,
  .small_size_mode_cluster = Inf,
  .maxT = 50,
  .sigma_string = "35\%",
  .diff_eps = 1e-06,
  verbose = FALSE
)
}
\arguments{
\item{truth_grouped_df}{data frame with filaments information (can be a 
grouped_df) to be assessed.}

\item{simulations_grouped_df}{grouped data frame with multiple simulated 
filaments information}

\item{data_column_names}{columns of \code{df_row_group} and 
\code{simulations_group_df} that correspond to the output space.}

\item{number_points}{the number of points the filament should be compressed 
to (if \code{Inf} then no compression is applied.}

\item{.change_radius}{boolean, if we should use the changing radius approach}

\item{.mode_cluster}{boolean, if we should use mode clustering}

\item{.to_simplex}{boolean, if we should project points on the unit simplex.}

\item{.use_frac}{boolean, if distances should be defined relative to a
scaling for the number of points in the path.}

\item{.small_size_mode_cluster}{int, the number of points the filament should 
be compressed to when estimating the mode clustering. We recommend a smaller
number to speed up the mode clustering algorithm.  (if Inf, then no change in 
size relative to \code{number_points})}

\item{.maxT}{int, max number of steps for mode clustering mean-shift 
algorithm}

\item{.sigma_string}{string, the quantile of the distance matrix to define
the sigma (e.g. \code{"30\%"})}

\item{.diff_eps}{float, the error allows between mode clustered final steps
to still be counted as the same mode.}

\item{verbose}{boolean, be verbose about progression}
}
\value{
list of information...
}
\description{
Mode clustering also is able to be on more course approximate of the 
filaments.
}
