% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conformal_prediction.R
\name{simulation_based_conformal3}
\alias{simulation_based_conformal3}
\title{global wrapper for simulation-based conformal score calculation that allows 
for mode clustering and changes of radius.}
\usage{
simulation_based_conformal3(
  truth_grouped_df,
  simulations_grouped_df,
  data_column_names = c("S", "I", "R"),
  number_points = 100,
  .change_radius = TRUE,
  .mode_cluster = TRUE,
  .to_simplex = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{truth_grouped_df}{data frame with filaments information (can be a 
grouped_df) to be assessed.}

\item{simulations_grouped_df}{grouped data frame with multiple simulated 
filaments information}

\item{data_column_names}{columns of \code{df_row_group} and 
\code{simulations_group_df} that correspond to the output space.}

\item{number_points}{the number of points the filament should be compressed 
to (if \code{Inf}) then no compression is applied.}

\item{.change_radius}{boolean, if we should use the changing radius approach}

\item{.mode_cluster}{boolean, if we should use mode clustering}

\item{.to_simplex}{boolean, if we should project points on the unit simplex.}

\item{verbose}{boolean, be verbose about progression}
}
\value{
data.frame with a row for each filament in \code{truth_grouped_df} 
with a second column \code{containment_val} which is the discrete 
simulation-based conformal score.
}
\description{
global wrapper for simulation-based conformal score calculation that allows 
for mode clustering and changes of radius.
}
