% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conformal_prediction.R
\name{simulation_based_conformal4}
\alias{simulation_based_conformal4}
\title{global wrapper for simulation-based conformal score calculation that allows 
for mode clustering and changes of radius.}
\usage{
simulation_based_conformal4(
  truth_grouped_df,
  simulations_grouped_df,
  smoothed_function,
  data_column_names = c("S", "I", "R"),
  number_points = Inf,
  .to_simplex = FALSE,
  .use_frac = TRUE,
  .small_size_mode_cluster = 50,
  .maxT = 50,
  .sigma_string = "35\%",
  .diff_eps = 1e-06,
  .eps = 1e-06,
  verbose = FALSE,
  return_min = FALSE
)
}
\arguments{
\item{truth_grouped_df}{data frame with filaments information (can be a 
grouped_df) to be assessed.}

\item{simulations_grouped_df}{grouped data frame with multiple simulated 
filaments information}

\item{smoothed_function}{a function for the smoothing approach for the varied
approach}

\item{data_column_names}{columns of \code{df_row_group} and 
\code{simulations_group_df} that correspond to the output space.}

\item{number_points}{the number of points the filament should be compressed 
to (if \code{Inf} then no compression is applied.}

\item{.to_simplex}{boolean, if we should project points on the unit simplex.}

\item{.use_frac}{boolean, if distances should be defined relative to a
scaling for the number of points in the path.}

\item{.small_size_mode_cluster}{int, the number of points the filament should 
be compressed to when estimating the mode clustering. We recommend a smaller
number to speed up the mode clustering algorithm.  (if Inf, then no change in 
size relative to \code{number_points})}

\item{.maxT}{int, max number of steps for mode clustering mean-shift 
algorithm}

\item{.sigma_string}{string, the quantile of the distance matrix to define
the sigma (e.g. \code{"30\%"})}

\item{.diff_eps}{float, the error allows between mode clustered final steps
to still be counted as the same mode.}

\item{.eps}{float, when to stop the model clustering (if maxT doesn't stop it
first)}

\item{verbose}{boolean, be verbose about progression}

\item{return_min}{boolean, if list of information returned is mimimum (for slurm)}
}
\value{
list of information...
}
\description{
Does all the 4 types of approaches (mode vs non mode clustering) and (
change in radius vs fixed radius)
}
