% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sis_data}
\alias{sis_data}
\title{Example SIS model  of class `dcm` from the \code{EpiModel} package.}
\format{
The object is a 'dcm', which is output from the EpiModel R package.
}
\usage{
sis_data
}
\description{
This is a deterministic SIS model where 
the infection parameter \eqn{\beta = 0.2} and the recovery/re-susceptibility
parameter is \eqn{\nu = 0.05}.  The initial number of susceptible is \eqn{S(0) = 499}
and the initial number of infectious is \eqn{I(0)=1}.  There are 100 steps.
For more details, see <<http://statnet.org/tut/BasicDCMs.html#sis_model_with_sensitivity_analyses>>.
}
\examples{
library(EpiModel)
data(sis_data)
plot(sis_data)
}
\keyword{datasets}
