% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sis_data_f}
\alias{sis_data_f}
\title{Example SIS model converted to aggregate format}
\format{
The object has the following columns
\describe{
\item{t}{time}
\item{orig_t}{original t in simulation}
\item{sim}{simulation number}
\item{X0}{number of Susceptible}
\item{X1}{number of Infectious}
}
}
\usage{
sis_data_f
}
\description{
This is a deterministic SIS model where 
the infection parameter \eqn{\beta = 0.2} and the recovery/re-susceptibility
parameter is \eqn{\nu = 0.05}.  The initial number of susceptible is \eqn{S(0) = 499}
and the initial number of infectious is \eqn{I(0)=1}.  There are 100 steps.
For more details, see <<http://statnet.org/tut/BasicDCMs.html#sis_model_with_sensitivity_analyses>>.  
We then use \code{fortify_aggregate()} to convert the data to output used with \code{EpiCompare} functions.
}
\examples{
library(EpiModel)
data(sis_data_f)
head(sis_data_f)
}
\keyword{datasets}
