% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{stat_sir}
\alias{stat_sir}
\title{SIR path visuals}
\usage{
stat_sir(mapping = NULL, data = NULL, geom = "path",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, data_type = c("raw", "fortified"), ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by
\code{\link[ggplot2:aes]{aes()}} or \code{\link[ggplot2:aes_]{aes_()}}.
If specified and \code{inherit.aes = TRUE} (the default), it is combined with
the default mapping at the top level of the plot. You must supply mapping if
there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three options:

If \code{NULL}, the default, the data is inherited from the plot data as
specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, will override the plot data.

A function will be called with a single argument, the plot data. The return
value must be a data.frame, and will be used as the layer data. A function
can be created from a formula (e.g. ~ head(.x, 10)).}

\item{geom}{Override the default connection between stat_sir() and
\code{\link[ggplot2:geom_path]{geom_path()}}.}

\item{position}{Position adjustment, either as a string, or the result of a
call to a position adjustment function}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with a
warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped. \code{FALSE}
 never includes, and \code{TRUE} always includes. It can also be a named
 logical vector to finely select the aesthetics to display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics, rather
than combining with them. This is most useful for helper functions that
define both data and aesthetics and shouldn't inherit behaviour from the}

\item{data_type}{default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}.
These are often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters to the
paired geom/stat.}
}
\description{
SIR path visuals
}
\section{Aesthetics}{

\code{stat_SirRaw()} understands the following aesthetics (required aesthetics are in bold):
\itemize{
  \item \strong{\code{y}}
  \item \strong{\code{z}}
  \item \code{group}
}
Learn more about setting these aesthetics in \code{vignette("ggplot2-specs")}.


\code{stat_SirFortified()} understands the following aesthetics (required aesthetics are in bold):
\itemize{
  \item \strong{\code{init_state}}
  \item \strong{\code{y}}
  \item \strong{\code{z}}
  \item \code{group}
}
Learn more about setting these aesthetics in \code{vignette("ggplot2-specs")}.
}

\examples{
library(ggplot2)
library(dplyr)


timeternR::hagelloch_raw \%>\%
  dplyr::filter(SEX \%in\% c("male", "female")) \%>\%
  ggplot(., aes(y = tI, z = tR, color = SEX)) +
    geom_path(stat = StatSirRaw) + ggtern::coord_tern() +
    labs(x = "S", y = "I", z = "R",
       color = "Gender")

timeternR::hagelloch_raw \%>\%
  dplyr::filter(SEX \%in\% c("male", "female")) \%>\%
  ggplot(., aes(y = tI, z = tR, color = SEX)) +
    stat_sir(geom = "path") + # note geom = "path" is the default
    ggtern::coord_tern() +
    labs(x = "S", y = "I", z = "R",
       color = "Gender")
}
