% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bands.R
\docType{data}
\name{StatConfBandConvexHull}
\alias{StatConfBandConvexHull}
\title{stat object for use in convex hull based stat_confidence_band and
geom_confidence_band}
\format{An object of class \code{StatConfBandConvexHull} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 3.}
\usage{
StatConfBandConvexHull
}
\description{
stat object for use in convex hull based stat_confidence_band and
geom_confidence_band
}
\keyword{datasets}
