% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bands.R
\docType{data}
\name{StatConfBandKDE}
\alias{StatConfBandKDE}
\title{stat object for use in kde based stat_confidence_band and
geom_confidence_band}
\format{An object of class \code{StatConfBandKDE} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 3.}
\usage{
StatConfBandKDE
}
\description{
stat object for use in kde based stat_confidence_band and
geom_confidence_band
}
\keyword{datasets}
