% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir-simulations.R
\name{state_change_inds}
\alias{state_change_inds}
\title{Determine which agents changed states}
\usage{
state_change_inds(new_states, current_states, type = "inf")
}
\arguments{
\item{new_states}{vector of size n_agents of 0/1/2}

\item{current_states}{vector of size n_agents of 0/1/2}

\item{type}{either "inf" for new infectious or "rec" for new recovered
return indices of which agents who were in state type are now in the next state}
}
\description{
Determine which agents changed states
}
