% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bands.R
\docType{data}
\name{StatPredBandConvexHull}
\alias{StatPredBandConvexHull}
\title{stat object for use in convex hull based stat_prediction_band and
geom_prediction_band}
\format{
An object of class \code{StatPredBandConvexHull} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 3.
}
\usage{
StatPredBandConvexHull
}
\description{
stat object for use in convex hull based stat_prediction_band and
geom_prediction_band
}
\keyword{datasets}
