% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bands.R
\docType{data}
\name{StatPredBandDeltaBall}
\alias{StatPredBandDeltaBall}
\title{stat object for use in delta_ball based stat_prediction_band and
geom_prediction_band}
\format{
An object of class \code{StatPredBandDeltaBall} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 4.
}
\usage{
StatPredBandDeltaBall
}
\description{
stat object for use in delta_ball based stat_prediction_band and
geom_prediction_band
}
\keyword{datasets}
