% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bands.R
\docType{data}
\name{StatPredBandKDE}
\alias{StatPredBandKDE}
\title{stat object for use in kde based stat_prediction_band and
geom_prediction_band}
\format{
An object of class \code{StatPredBandKDE} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 4.
}
\usage{
StatPredBandKDE
}
\description{
stat object for use in kde based stat_prediction_band and
geom_prediction_band
}
\keyword{datasets}
