% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bands.R
\docType{data}
\name{StatPredBandSpherical}
\alias{StatPredBandSpherical}
\title{stat object for use in spherical ball based stat_prediction_band and
geom_prediction_band}
\format{
An object of class \code{StatPredBandSpherical} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 4.
}
\usage{
StatPredBandSpherical
}
\description{
stat object for use in spherical ball based stat_prediction_band and
geom_prediction_band
}
\keyword{datasets}
