% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bands.R
\name{step_along}
\alias{step_along}
\title{step along path}
\usage{
step_along(start_point, angle, distance)
}
\arguments{
\item{start_point}{(x,y) starting point}

\item{angle}{angle away from point (in radians)}

\item{distance}{distance away from point for the new point}
}
\value{
next point along the correct angle and distance
}
\description{
step along path
}
