% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance-between.R
\name{step_along_angle}
\alias{step_along_angle}
\title{step along 2d path with angle}
\usage{
step_along_angle(start_point, angle, distance)
}
\arguments{
\item{start_point}{(x,y) starting point}

\item{angle}{angle away from point (in radians)}

\item{distance}{distance away from point for the new point}
}
\value{
next point along the correct angle and distance
}
\description{
step along 2d path with angle
}
