% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance-between.R
\name{step_along_direction}
\alias{step_along_direction}
\title{step along path with direction}
\usage{
step_along_direction(start_point, direction, distance)
}
\arguments{
\item{start_point}{(d vector) starting point}

\item{direction}{path away from starting point}

\item{distance}{distance away from point for the new point}
}
\value{
next point along the correct angle and distance
}
\description{
step along path with direction
}
