% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-dist-mat.R
\name{[.tidy_dist_mat}
\alias{[.tidy_dist_mat}
\title{Extract a part of a \code{tidy_dist_mat} object}
\usage{
\method{[}{tidy_dist_mat}(x, i, j = i)
}
\arguments{
\item{x}{\code{tidy_dist_mat} object}

\item{i}{indices specifying elements to extract or replace. Indices are 
numeric or data frames related to \code{x}'s row information.}

\item{j}{indices specifying elements to extract or replace. Indices are 
numeric or data frames related to \code{x}'s column information. If no 
\code{j} is provided, then \code{j=i} and a square matrix is returned (if 
\code{x}'s structure follows correct assumptions).}
}
\value{
a new \code{tidy_dist_mat} object
}
\description{
Extract a part of a \code{tidy_dist_mat} object
}
