% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-dist-mat.R
\name{tail.tidy_dist_mat}
\alias{tail.tidy_dist_mat}
\title{Return the Last Parts of a \code{tidy_dist_mat} (symmetric grab)}
\usage{
\method{tail}{tidy_dist_mat}(x, n = 6L, ...)
}
\arguments{
\item{x}{symmetric (in shape) \code{tidy_dist_mat} object}

\item{n}{an integer of value bounded by the maximum rows \code{x}.}

\item{...}{arguments to be passed to or from other methods (currently none
expected / used).}
}
\value{
\code{x[(nrow(x)-n):nrow(x)]}
}
\description{
Return the Last Parts of a \code{tidy_dist_mat} (symmetric grab)
}
