% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_peakfile.R
\name{tidy_peakfile}
\alias{tidy_peakfile}
\title{Tidy peakfiles in GRanges}
\usage{
tidy_peakfile(peaklist, blacklist)
}
\arguments{
\item{peaklist}{A list of peak files as GRanges object. If more than one, objects in lists using `list()`.}

\item{blacklist}{Peakfile specifying blacklisted regions as GRanges object.}
}
\value{
list of GRanges object
}
\description{
Tidy peakfiles in GRanges
}
\examples{
library(EpiCompare)
data("encode_H3K27ac") # example dataset as GRanges object
data("CnT_H3K27ac") # example dataset as GRanges object
data("hg19_blacklist") # blacklist region for hg19 genome

tidy_peakfile(peaklist = list(encode_H3K27ac, CnT_H3K27ac),
                     blacklist = hg19_blacklist)

}
