% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tidyr_new_interface}
\alias{tidyr_new_interface}
\title{logic to check if tidyverse (and tidyr specifically is up to version 1.0)}
\usage{
tidyr_new_interface()
}
\value{
logical value (boolean)
}
\description{
logic to check if tidyverse (and tidyr specifically is up to version 1.0)
}
