% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeternR-package.R
\docType{package}
\name{timeternR-package}
\alias{timeternR}
\alias{timeternR-package}
\title{timeternR: functional based data visualized with ternary plots in the simplex.}
\description{
Nothing yet...
}
\author{
\strong{Maintainer}: Shannon Gallagher \email{sgallagh@stat.cmu.edu}

Authors:
\itemize{
  \item Benjamin LeRoy \email{bpleroy@stat.cmu.edu}
}

}
\keyword{internal}
