% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/containment-and-band-creation.R
\name{to_lower_simplex}
\alias{to_lower_simplex}
\title{move df to simplex representation}
\usage{
to_lower_simplex(df, A = NULL)
}
\arguments{
\item{df}{an n x p data.frame, all scalar columns}

\item{A}{default is \code{NULL}, if so, we create A from
\code{simplex_project_mat}}
}
\value{
updated dat.frame (n x k), where k should be p-1, but that's assuming
  A is made from \code{simplex_project_mat}
}
\description{
This function takes a data.frame, moves it to unit simplex representation and
then projects the points onto a lower representation space (using \code{A}).
}
\examples{
center <- data.frame(x = 1, y = 1, z = 1)
to_lower_simplex(center) # 0,0 # center

side <- data.frame(x = 1, y = 0, z =0)
to_lower_simplex(side) #(1,0)
}
