% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition-matrix-to-function.R
\name{trans_mat_to_D_fxn}
\alias{trans_mat_to_D_fxn}
\title{Take a character formula transition matrix and turn it into a function}
\usage{
trans_mat_to_D_fxn(D_mat, K, n_pars)
}
\arguments{
\item{D_mat}{a KxK matrix of character transitions.  See details}

\item{K}{number of states}

\item{n_pars}{number of parameters}
}
\value{
a list of length \eqn{K^2} where entry K(i-1) + j for i, j <= K is a function.  See details.
}
\description{
Take a character formula transition matrix and turn it into a function
}
\details{
This function dynamically creates a list of functions that turn the
 character transition formulas provided by the user in D_mat into \eqn{K^2}
  functions which take in arguments X0, ..., X{K-1)}, par1, par2, ..., par{n_pars},
 N, and t and output the numeric transition rate of an individual
  currently in state \eqn{i} to move to state \eqn{j}.
}
