% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{U_sims_tidy}
\alias{U_sims_tidy}
\title{U_sim_tidy Example output from SIR simulations in a data.frame format}
\format{An 9400 x 5 data frame where each row is an agent for a given simulation and the columns are
\describe{
  \item{sim}{Simulation number between 1 and 50}
  \item{individual}{This is the ID of the individual}
  \item{init_state}{The initial state of the agent at time 0.  It is 0/1/2 for susceptible, infectious, recovered, respectively.}
  \item{SMax}{maximum time individual was suspectable (S)}
  \item{IMax}{maximum time individual was suspectable (I)}
}}
\usage{
U_sims_tidy
}
\description{
This simulated data is an example of the output from the function \code{simulate_SIR_agents()} when \code{output_format = "data.frame"}.  It is a 9400 by 5 data.frame where each row corresponds to the U statistic of a given agent and simulation.
Each row is asssociated with 1 agent and contains
is a "sufficient" statistic for each agent's infection.  Each agent's
infection is uniquely identified by an initial state, max time before
infection (or time T-1), and max time before recovery (or time T-1). Note time T-1
is the final recorded time in the SIR model (in this case T = 50).
}
\examples{
## show first simulation of first 10 agents
U_sims[1, , 1:10]
}
\keyword{datasets}
