% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thesis_data.R
\name{uniform_check_wrapper}
\alias{uniform_check_wrapper}
\title{test how close conformal scores are to discrete uniform distribution
across functions}
\usage{
uniform_check_wrapper(
  info_x,
  sim_list,
  sim_names,
  info_x_test,
  pseudo_density_df,
  grouping_df_list,
  named_smooth_function_list,
  tdm_sim,
  mm_delta,
  dist_between_array,
  verbose = T
)
}
\arguments{
\item{info_x}{}

\item{sim_list}{}

\item{sim_names}{}

\item{info_x_test}{}

\item{pseudo_density_df}{}

\item{grouping_df_list}{}

\item{named_smooth_function_list}{}

\item{tdm_sim}{}

\item{mm_delta}{}
}
\value{

}
\description{
test how close conformal scores are to discrete uniform distribution
across functions
}
