% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{update_approved_layers}
\alias{update_approved_layers}
\title{Updates approved layers of ggtern}
\usage{
update_approved_layers(stat_names = NULL, geom_names = NULL)
}
\arguments{
\item{stat_names}{named vector of new stats.}

\item{geom_names}{named vector of new geoms}
}
\description{
Updates approved layers of ggtern
}
\examples{
.newstat = c(sir_raw       = "StatSirRaw",
             sir_fortified = "StatSirFortified")
.newgeom = NULL
#update_approved_layers(stat_name = .newstat, geom_name = .newgeom)
}
