% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thesis_data.R
\name{update_tm_smooth}
\alias{update_tm_smooth}
\title{update a tm_radius matrix by smoothing about minim covering radi values}
\usage{
update_tm_smooth(tm_radius, func)
}
\arguments{
\item{tm_radius}{original list of lists, each containing a 
\code{min_cover_vec} and a \code{dist_mat} to be updated.}

\item{func}{smoothing function that creates a  function that takes in 
\code{x} and \code{y} and returns a smoothed \code{y_hat} for the x values.}
}
\value{
updated \code{tm_radius}
}
\description{
update a tm_radius matrix by smoothing about minim covering radi values
}
