% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{UtoX_SIR}
\alias{UtoX_SIR}
\title{Convert agent information to SIR format}
\usage{
UtoX_SIR(U, max_time = NULL, ind = NULL)
}
\arguments{
\item{U}{data frame, with the following format
\describe{
  \item{init_state}{Initial state for individual (at time t = 0). For the
  states, 0 = S, 1 = I, 2 = R.}
  \item{max_time_S}{maximum time individual was suspectable (S)}
  \item{max_time_I}{maximum time individual was infected (I)}
}}

\item{max_time}{integer, max length of outbreak (default NULL)}

\item{ind}{integer vector which columns match up with the columns described
above (default NULL)}
}
\value{
\code{sir_out} data frame, with columns
\describe{
  \item{t}{time since outbreak}
  \item{S}{Number of individuals suspectable}
  \item{I}{Number of individuals infected}
  \item{R}{Number of individuals in recovery}
}
}
\description{
Convert agent information to SIR format
}
\examples{
sir_out <- UtoX_SIR(timeternR::hagelloch_agents)
assertthat::are_equal(sir_out, timeternR::hagelloch_sir)
}
