% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_parameters.R
\name{validate_parameters}
\alias{validate_parameters}
\title{Validate input parameters}
\usage{
validate_parameters(peaklist, namelist, reference, stat_plot)
}
\value{
No return
}
\description{
Validate input parameters
}
\keyword{internal}
