% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-dist-mat.R
\name{which_index}
\alias{which_index}
\alias{which_index.tidy_dist_mat}
\alias{which_not_index}
\alias{which_not_index.tidy_dist_mat}
\title{find the index of a \code{tidy_dist_mat} relative to a data.frame}
\usage{
which_index(x, index, margin = 1)

\method{which_index}{tidy_dist_mat}(x, index, margin = 1)

which_not_index(x, index, margin = 1)

\method{which_not_index}{tidy_dist_mat}(x, index, margin = 1)
}
\arguments{
\item{x}{\code{tidy_dist_mat} object}

\item{index}{data.frame index associated with \code{x}'s colnames or
rownames (related to \code{margin}). If \code{which_not_index} then this 
is which associated information *not* to include.}

\item{margin}{scalar representing if \code{index} is related to rows or 
columns of \code{x}. 1 = row, 2 = column.}
}
\value{

}
\description{
probably more of an internal function
}
