% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_example_peaks.R
\name{write_example_peaks}
\alias{write_example_peaks}
\title{Write example peaks}
\usage{
write_example_peaks(
  dir = file.path(tempdir(), "processed_results"),
  datasets = c("encode_H3K27ac", "CnT_H3K27ac", "CnR_H3K27ac")
)
}
\arguments{
\item{dir}{Directory to save peak files to.}

\item{datasets}{Example datasets from \pkg{EpiCompare} to write.}
}
\value{
Named vector of paths to saved peak files.
}
\description{
Write example peaks datasets to disk.
}
\examples{
save_paths <- EpiCompare::write_example_peaks()
}
