% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent-to-aggregate.R
\name{agents_to_aggregate}
\alias{agents_to_aggregate}
\title{generalized method to convert raw agent based data to aggregate data}
\usage{
agents_to_aggregate(
  agents,
  states,
  death = NULL,
  birth = NULL,
  min_max_time = c(0, NA)
)
}
\arguments{
\item{agents}{data frame style object (currently either of class
\code{data.frame} or \code{grouped_df})}

\item{states}{time entered state. Do not include column for original state.
These need to be ordered, for example: for an SIR model, with columns
"\code{tI}" and "\code{tR}" expressing the time the individual became
infected and recovered (respectively), we want
"\code{states = c("tI", "tR")}".}

\item{death}{string for column with death time information (default
\code{NULL})}

\item{birth}{string for column with birth time information (default
\code{NULL})}

\item{min_max_time}{vector (length 2) of minimum and maximum integer time,
the second value can be \code{NA} - and if so, we estimate maximum time from
the data.}
}
\value{
dataset with aggregated information, we label classes
"\code{X\{i\}}" for i in \code{0:(length(states))}. Potentially
calculated per group of a \code{grouped_df}.
}
\description{
generalized method to convert raw agent based data to aggregate data
}
\examples{
# see ?agent_to_aggregate.data.frame or ?agent_to_aggregate.grouped_df
# for examples
}
