% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_genome_build.R
\name{check_genome_build}
\alias{check_genome_build}
\title{Check genome build}
\usage{
check_genome_build(genome_build)
}
\arguments{
\item{genome_build}{Genome build name.}
}
\value{
txdb
}
\description{
Check that the genome build is valid and 
require specific reference datasets to be installed.
}
\keyword{internal}
