% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_list_names.R
\name{check_list_names}
\alias{check_list_names}
\title{Check peaklist is named}
\usage{
check_list_names(peaklist, default_prefix = "sample")
}
\arguments{
\item{peaklist}{A list of peak files as GRanges object.}

\item{default_prefix}{Default prefix to use when creating names
for \code{peaklist}.}
}
\value{
named peaklist
}
\description{
This function checks whether the peaklist is named. 
If not, default file names are assigned.
}
\keyword{internal}
