% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent-to-aggregate.R
\name{check_min_max_time}
\alias{check_min_max_time}
\title{min_max_time vector check}
\usage{
check_min_max_time(min_max_time)
}
\arguments{
\item{min_max_time}{min_max_time}
}
\value{
error or \code{TRUE} if min_max_time vector meets assumptions
}
\description{
this is an internal function
}
