% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent_to_agg.R
\name{check_ordered}
\alias{check_ordered}
\title{checks if states within data frame are ordered as inputted (<=)}
\usage{
check_ordered(df, states, assert_error = TRUE)
}
\arguments{
\item{df}{data frame with individual agent information}

\item{states}{time entered state, in expected order}

\item{assert_error}{boolean if we should raise error if ordering assumption
is violated.}
}
\value{
depends, see details
}
\description{
checks if states within data frame are ordered as inputted (<=)
}
\details{
This function returns either 1) TRUE, if the states are ordered, 2)
an error is observed (when assert_error = TRUE), or 3) a list of 2 data
frames, the first containing information for each individual (their observed
ordering and if they broke the provided ordering), and the second containing
summarization on the number of individuals in each of the ordering "classes".
}
