% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chromHMM_annotation_K562}
\alias{chromHMM_annotation_K562}
\title{ChromHMM annotation file subset}
\format{
An object of class \code{CompressedGRangesList} of length 15.
}
\source{
The code to prepare the .Rda file from the raw peak file is:
\code{
chrHMM_url <- "http://hgdownload.cse.ucsc.edu/goldenPath/hg19/encodeDCC/wgEncodeBroadHmm/wgEncodeBroadHmmK562HMM.bed.gz"
chrHMM <- genomation::readBed(chrHMM_url)
chr1_seqnames <- subset(chrHMM@seqnames, factor(chrHMM@seqnames) == "chr1") # subset chr1 seqnames
chr1_ranges <- chrHMM@ranges[1:67294] # subset chr1 ranges
chr1_strand <- chrHMM@strand[1:67294] # subset chr1 strand
chr1_elementMetadata <- chrHMM@elementMetadata[1:67294,] # subset chr1 element metadata
chrHMM@seqnames <- chr1_seqnames
chrHMM@ranges <- chr1_ranges
chrHMM@strand <- chr1_strand
chrHMM@elementMetadata <- chr1_elementMetadata
chromHMM_annotation_K562 <- GenomicRanges::split(chrHMM, chrHMM$name, drop = TRUE)
usethis::use_data(chromHMM_annotation_K562, overwrite = TRUE)
}
}
\usage{
chromHMM_annotation_K562
}
\description{
ChromHMM annotation for human K562 cell-line.
Obtained from \url{http://hgdownload.cse.ucsc.edu/goldenPath/hg19/encodeDCC/wgEncodeBroadHmm/}
The annotation was subset for chromosome 1 only to reduce the size of data.
}
\keyword{datasets}
