% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CnT_H3K27ac}
\alias{CnT_H3K27ac}
\title{Example CUT&Tag peak file}
\format{
An object of class \code{GRanges} of length 13456.
}
\source{
The code to prepare the .Rda file from the raw peak file is:
\code{
CnT_H3K27ac <- ChIPseeker::readPeakFile("./data/CnT_H3K27ac_MACS2.bed", as = "GRanges")
usethis::use_data(CnT_H3K27ac)
}
}
\usage{
CnT_H3K27ac
}
\description{
Human H3K27ac peak file generated with CUT&Tag using K562 cell-line from Kaya-Okur et al., (2019).
Raw peak file (.BED) was obtained from GEO (sample accession: SRR8383507).
Peak calling was performed by Leyla Abbasova using MACS2.
The peak file was then processed into GRanges object.
}
\keyword{datasets}
