% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-dist-mat.R
\name{colnames<-.tidy_dist_mat}
\alias{colnames<-.tidy_dist_mat}
\title{assign colnames}
\usage{
\method{colnames}{tidy_dist_mat}(x) <- value
}
\arguments{
\item{x}{tidy_dist_mat}

\item{value}{updated rownames}
}
\description{
\url{https://stackoverflow.com/questions/11285496/r-cmd-check-warning-functions-methods-with-usage-in-documentation-object-bu}
}
