% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bands.R
\name{delta_structure}
\alias{delta_structure}
\title{Performs delta ball approach}
\usage{
delta_structure(data_deep_points)
}
\arguments{
\item{data_deep_points}{data deep points from depth function}
}
\value{
\describe{
\item{structure}{data frame of non-ordered lines of contour}
\item{delta}{optimal delta for covering}
}
}
\description{
Performs delta ball approach
}
