% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance-between.R
\name{dist_between_paths}
\alias{dist_between_paths}
\title{Distance between points}
\usage{
dist_between_paths(data_df_p1, data_df_p2)
}
\arguments{
\item{data_df_p1}{(n x 2) data.frame, each row is a pair of values (x, y)}

\item{data_df_p2}{(n x 2) second data.frame to be compared, each row is a
pair of values (x, y)}
}
\value{
vector of n distances between each order point in the two data frames
}
\description{
Calculates the distance between each point of every path.
}
