% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance-between.R
\name{dist_matrix_innersq_direction}
\alias{dist_matrix_innersq_direction}
\title{Calculates the distance matrix between a set of paths (Euclidean based). This
is actually d^2}
\usage{
dist_matrix_innersq_direction(path_list, position = NULL, verbose = FALSE)
}
\arguments{
\item{path_list}{list of paths (data frames) - need to have the same num rows}

\item{position}{column index of (x,y) euclidean coords.}

\item{verbose}{boolean logic if should have print outs while computing
distance matrix}
}
\value{
distance matrix of dimension n x n
}
\description{
Calculates the distance matrix between a set of paths (Euclidean based). This
is actually d^2
}
