% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance-between.R
\name{dist_matrix_innersq_direction}
\alias{dist_matrix_innersq_direction}
\alias{dist_matrix_innersq_direction.list}
\alias{dist_matrix_innersq_direction.data.frame}
\alias{dist_matrix_innersq_direction.grouped_df}
\title{Calculates the distance matrix between a set of paths (Euclidean based). This
is actually d^2}
\usage{
dist_matrix_innersq_direction(
  x,
  position = NULL,
  verbose = FALSE,
  tdm_out = FALSE
)

\method{dist_matrix_innersq_direction}{list}(
  x,
  position = NULL,
  verbose = FALSE,
  tdm_out = FALSE
)

\method{dist_matrix_innersq_direction}{data.frame}(
  x,
  position = NULL,
  verbose = FALSE,
  tdm_out = FALSE
)

\method{dist_matrix_innersq_direction}{grouped_df}(
  x,
  position = NULL,
  verbose = FALSE,
  tdm_out = FALSE
)
}
\arguments{
\item{x}{list of paths (data frames) - need to have the same num rows or a nested_df}

\item{position}{column index of (x,y,...) euclidean coords.}

\item{verbose}{boolean logic if should have print outs while computing
distance matrix}

\item{tdm_out}{boolean logic if we should return a \code{tidy_dist_mat} 
object or just a matrix.}
}
\value{
distance matrix of dimension n x n or data.frame - DESCRIBE
}
\description{
Calculates the distance matrix between a set of paths (Euclidean based). This
is actually d^2
}
