% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance-depth.R
\name{distance_depth_function}
\alias{distance_depth_function}
\alias{distance_depth_function.matrix}
\alias{distance_depth_function.tidy_dist_mat}
\title{Geenens & Nieto-Reyes functional distance-based depth}
\usage{
distance_depth_function(dist_mat, df_out = "auto")

\method{distance_depth_function}{matrix}(dist_mat, df_out = F)

\method{distance_depth_function}{tidy_dist_mat}(dist_mat, df_out = T)
}
\arguments{
\item{dist_mat}{a n x n square positive symmetric matrix or a tidy_dist_mat}

\item{df_out}{indicates if one should return a data.frame our a vector,
by default returns data.frame if dist_mat is a tidy_dist_mat, and a vector
if dist_mat is a matrix.}
}
\value{
depth vector length n with depth values associated with indices in
  \code{dist_mat} or a data.frame with a column called \code{depth}
}
\description{
Calculates a global distance-based depth vector using a distance matrix.
Specifically we use Geenens & Nieto-Reyes's global distance-based depth
defined as:
}
\details{
\eqn{DD(x, \hat{P}) = 1/(n choose 2) \cdot \sum_{i!=j} I(d(X_i, X_j) >
max(d(X_i,x), d(X_j,x)))}


This function (renamed as \code{depth_function}) is shared with
\pkg{TCpredictionbands} on github:
\href{https://github.com/Mr8ND/TC-prediction-bands/tree/master/TCpredictionbands}{TCpredictionbands}.
}
\examples{
dist_mat <- matrix(c(0,   1, 1.5,
                     1,   0, 2,
                     1.5, 2, 0   ),
                   nrow = 3,
                   byrow = TRUE)

dd_vec <- distance_depth_function(dist_mat) # c(1,0,0)

tidy_dm <- tidy_dist_mat(dist_mat)
dd_df <- distance_depth_function(tidy_dm) # depth = c(1,0,0)
}
