% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{encode_H3K27ac}
\alias{encode_H3K27ac}
\title{Example ChIP-seq peak file}
\format{
An object of class \code{GRanges} of length 5142.
}
\source{
The code to prepare the .Rda file from the raw peak file is:
\code{
encode_H3K27ac <- ChIPseeker::readPeakFile("./data/ENCODE_H3K27ac.bed", as = "GRanges")
encode_H3K27ac <- encode_H3K27ac[encode_H3K27ac@seqnames == "chr1"]
usethis::use_data(encode_H3K27ac)
}
}
\usage{
encode_H3K27ac
}
\description{
Human H3K27ac peak file generated with ChIP-seq using K562 cell-line.
Raw peak file (.BED) was obtained from ENCODE project (data accession: ENCFF044JNJ).
The BED file was then processed into GRanges object.
Peaks located on chromosome 1 were subsetted to reduce the dataset size.
}
\keyword{datasets}
