% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiCompare.R
\name{EpiCompare}
\alias{EpiCompare}
\title{Compare Epigenetic Datasets}
\usage{
EpiCompare(peakfile1_path, peakfile2_path, outpath)
}
\description{
This function calculates percentage of overlapping peaks in two different peak files.
It also performs ChromHMM on individual files, overlapping and unique peaks.
}
