% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiCompare.R
\name{EpiCompare}
\alias{EpiCompare}
\title{Compare Two Epigenetic Datasets}
\usage{
EpiCompare(peakfile1_path, peakfile2_path, outpath)
}
\arguments{
\item{peakfile1_path}{Path to your peakfile 1}

\item{peakfile2_path}{Path to your peakfile 2}

\item{outpath}{Path to where you want the knitted HTML report}
}
\value{
a knitted HTML report containing results of the analysis
}
\description{
This function calculates the percentage of overlapping peaks in two peak files.
It also performs ChromHMM on individual peak files, overlapping and unique peaks.
The output of this function is a knitted HTML file, which will be saved in the specified outpath.
}
\examples{
EpiCompare(peakfile1_path = "path/to/peakfile1",
           peakfile2_path = "path/to/peakfile2",
           outpath = "path/for/output/report" )

}
