% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiCompare.R
\name{EpiCompare}
\alias{EpiCompare}
\title{Compare Two Epigenetic Datasets}
\usage{
EpiCompare(peakfile1, peakfile1_name, peakfile2, peakfile2_name, outpath)
}
\arguments{
\item{peakfile1}{Peakfile1 as GRanges object}

\item{peakfile1_name}{Name of peakfile1 as strings (inserted in single/double quotation marks)}

\item{peakfile2}{Peakfile2 as GRanges object}

\item{peakfile2_name}{Name of peakfile2 as strings (inserted in single/double quotation marks)}

\item{outpath}{Path to where you want the knitted HTML report}
}
\value{
a knitted HTML report containing results of the analysis
}
\description{
This function calculates the percentage of overlapping peaks in two peak files.
It also performs ChromHMM on individual peak files, overlapping and unique peaks.
The output of this function is a knitted HTML file, which will be saved in the specified outpath.
}
\examples{
library(EpiCompare)
data(encode_H3K27ac) # example dataset as GRanges object
data(CnT_H3K27ac) # example dataset as GRanges object

EpiCompare(peakfile1 = encode_H3K27ac,
           peakfile1_name = "ENCODE",
           peakfile2 = CnT_H3K27ac,
           peakfile2_name = "CUT&Tag",
           outpath = "./EpiCompare.html" )

}
