% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiCompare.R
\name{EpiCompare}
\alias{EpiCompare}
\title{Compare epigenetic datasets}
\usage{
EpiCompare(
  peakfiles,
  blacklist,
  picard_files = NULL,
  reference = NULL,
  stat_plot = FALSE,
  chrmHMM_plot = FALSE,
  chipseeker_plot = FALSE,
  enrichment_plot = FALSE,
  save_output = FALSE,
  output_dir
)
}
\arguments{
\item{peakfiles}{A list of peak files as GRanges object.
Files must be listed using `list()` and named using `names()`
If not named, default file names will be assigned.}

\item{blacklist}{A GRanges object containing blacklisted regions.}

\item{picard_files}{A list of summary metrics output from Picard.
Files must be in data.frame format and listed using `list()` and named using `names()`.
To import Picard duplication metrics (.txt file) into R as data frame,
use `picard <- read.table("/path/to/picard/output", header = TRUE, fill = TRUE)`.}

\item{reference}{A reference peak file as GRanges object.
If a reference is specified, it enables two analyses: (1) plot showing statistical significance of overlapping/non-overlapping peaks;
and (2) ChromHMM of overlapping/non-overlapping peaks. Please ensure that reference file is named using `names()`}

\item{stat_plot}{Default FALSE. If TRUE, the function creates a plot showing the statistical significance of
overlapping/non-overlapping peaks. Reference peak file must be provided.}

\item{chrmHMM_plot}{Default FALSE. If TRUE, the function outputs ChromHMM heatmap of individual peak files.
If a reference peak file is provided, ChromHMM annotation of overlapping and non-overlapping peaks is also provided.}

\item{chipseeker_plot}{Default FALSE. If TRUE, the report includes a barplot of ChIPseeker annotation of peak files.}

\item{enrichment_plot}{Default FALSE. If TRUE, the report includes dotplots of KEGG and GO enrichment analysis of peak files.}

\item{save_output}{Default FALSE. If TRUE, all outputs (tables and plots) of the analysis will be saved in a folder (EpiCompare_file).}

\item{output_dir}{Path to where output HTML file should be saved.}
}
\value{
An HTML report
}
\description{
This function compares different epigenetic datasets and performs various functional analyses.
The function outputs an HTML report containing results from the analysis.
The report is mainly divided into three areas: (1) Peakfile information, (2) Overlapping peaks and (3) Functional annotations.
}
\examples{
library(EpiCompare)
data("encode_H3K27ac") # example dataset as GRanges object
data("CnT_H3K27ac") # example dataset as GRanges object
data("CnR_H3K27ac") # example dataset as GRanges object
data("hg19_blacklist") # example blacklist dataset as GRanges object

peaks <- list(CnR_H3K27ac, CnT_H3K27ac) # create list
names(peaks) <- c("CnR", "CnT") # set names

EpiCompare(peakfiles = peaks,
           blacklist = hg19_blacklist,
           reference = encode_H3K27ac,
           stat_plot = TRUE,
           chrmHMM_plot = TRUE,
           chipseeker_plot = TRUE,
           enrichment_plot = TRUE,
           save_output = FALSE,
           output_dir = tempdir())

}
