% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiCompare.R
\name{EpiCompare}
\alias{EpiCompare}
\title{Compare epigenetic datasets}
\usage{
EpiCompare(
  peakfiles,
  names,
  blacklist,
  picard = NULL,
  picard_names = NULL,
  reference = NULL,
  stat_plot = FALSE,
  chrmHMM_plot = FALSE,
  save_output = FALSE,
  output_dir
)
}
\arguments{
\item{peakfiles}{A list of peak files as GRanges object. Files must be listed using `list()`}

\item{names}{A list of file names in the same order as the list of peak files.
Names must be specified using `c()`}

\item{blacklist}{A GRanges object containing blacklisted regions.}

\item{picard}{A list of summary metrics output from Picard.}

\item{picard_names}{A list of names in the same order as the list of Picard summary output.}

\item{reference}{A reference peak file as GRanges object.
If a reference is specified, it enables two analyses: (1) plot showing statistical significance of overlapping/non-overlapping peaks;
and (2) ChromHMM of overlapping/non-overlapping peaks}

\item{stat_plot}{Default FALSE. If TRUE, the function creates a plot showing the statistical significance of
overlapping/non-overlapping peaks. Reference peak file must be provided.}

\item{chrmHMM_plot}{Default FALSE. If TRUE, the function outputs ChromHMM heatmap of individual peak files.
If a reference peak file is provided, ChromHMM annotation of overlapping and non-overlapping peaks is also provided.}

\item{save_output}{Default FALSE. If TRUE, all outputs (tables and plots) of the analysis will be saved in a folder (EpiCompare_file).}

\item{output_dir}{Path to where output HTML file should be saved.}
}
\value{
An HTML report
}
\description{
This function compares different epigenetic datasets and performs various functional analyses.
The function outputs an HTML report containing results from the analysis.
The report is mainly divided into three areas: (1) Peakfile information, (2) Overlapping peaks and (3) Functional annotations.
}
\examples{
library(EpiCompare)
data("encode_H3K27ac") # example dataset as GRanges object
data("CnT_H3K27ac") # example dataset as GRanges object
data("CnR_H3K27ac") # example dataset as GRanges object
data("hg19_blacklist") # example blacklist dataset as GRanges object
data("chromHMM_annotation_K562") # example of ChromHMM annotation file

EpiCompare(peakfiles = list(CnR_H3K27ac, CnT_H3K27ac),
           names = c("CnR", "CnT"),
           blacklist = hg19_blacklist,
           reference = encode_H3K27ac,
           stat_plot = TRUE,
           chrmHMM_plot = TRUE,
           save_output = TRUE,
           output_dir = "./")

}
