% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent-to-aggregate.R, R/agent_to_agg.R
\name{expanding_info}
\alias{expanding_info}
\title{add class and time placeholder with zero individuals to data frame}
\usage{
expanding_info(df, t_min, t_max, K)

expanding_info(df, t_min, t_max, K)
}
\arguments{
\item{df}{data frame with \code{state}, \code{t} and \code{count} columns
which provide information on the number of agents that became said
\code{state} at time \code{t}.}

\item{t_min}{minimum integer time}

\item{t_max}{maximum integer time}

\item{K}{number of stages - 1 (e.g.: SIR has K = 2)}
}
\value{
expanded data frame

expanded data frame
}
\description{
this is an internal function (so t_min, t_max, K should be correct)

this is an internal function (so t_min, t_max, K should be correct)
}
