% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fortify_agents}
\alias{fortify_agents}
\title{fortify agent data frame with columns when individual stops being suspectable
and stops being infected (as well as initial state).}
\usage{
fortify_agents(raw_df, time_col = c("tI", "tR"),
  T = ceiling(max(raw_df[, time_col])) + 5)
}
\arguments{
\item{raw_df}{data frame, agent based data frame}

\item{time_col}{length 2 string vector, column names recording when
individual is infected and when they enter the recovery stage}

\item{T}{int, maximum time for infection process}
}
\value{
\code{fortified_df} data frame, the \code{raw_df} plus three
additional columns:
\describe{
  \item{init_state}{Initial state for individual (at time t = 0). For the
  states, 0 = S, 1 = I, 2 = R.}
  \item{max_time_S}{maximum time individual was suspectable (S)}
  \item{max_time_I}{maximum time individual was infected (I)}
}
}
\description{
fortify agent data frame with columns when individual stops being suspectable
and stops being infected (as well as initial state).
}
\examples{
fortify_df <- fortify_agents(timeternR::hagelloch_raw,
                             time_col = c("tI","tR"),
                             T = 90)
assertthat::are_equal(fortify_df[,(ncol(fortify_df) - 2):ncol(fortify_df)],
                      as.data.frame(t(timeternR::hagelloch_agents)))
}
