% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify-aggregate-epimodel-icm.R
\name{fortify_aggregate.icm}
\alias{fortify_aggregate.icm}
\title{Take external aggregate data and put it in a format used in this package}
\usage{
\method{fortify_aggregate}{icm}(
  data,
  states = NULL,
  enquo_states = NULL,
  package_source = NULL
)
}
\arguments{
\item{data}{output from external source package.  See details}

\item{states}{names of states we want aggregate totals of at each time}

\item{enquo_states}{vector of quoted names}

\item{package_source}{optional argument to include the package from which the
output is derived from, which helps with the fortify function when outputs
are of generic classes such as list or data.frame}
}
\value{
a data frame with the following columns
\describe{
\item{t}{time}
\item{Xk}{columns X0, ..., X_K. which are numeric}
}
}
\description{
Take external aggregate data and put it in a format used in this package
}
\details{
This function converts external data sources (we currently support
  output from the EpiModel and pomp R packages), which is already aggregated
  and puts it in a format that can be used by our exploring functions.
}
\examples{
## For icm
out <- fortify_aggregate(EpiModel_icm)
head(out)
}
