% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify_data.R
\name{fortify_EpiModel}
\alias{fortify_EpiModel}
\title{Generic method that takes in data from the \code{EpiModel} package and puts it in aggregate, SIR format}
\usage{
fortify_EpiModel(data)
}
\arguments{
\item{data}{Output from a EpiModel simulation of class \code{icm} or \code{dcm}}
}
\value{
data frame with the following columns
\describe{
\item{t}{the time}
\item{S}{number of Susceptibles}
\item{I}{number of Infectious}
\item{R}{number of Recovered}
\item{sim}{simulation number (factor variable) (optional column)}
}
}
\description{
Generic method that takes in data from the \code{EpiModel} package and puts it in aggregate, SIR format
}
\examples{
## For icm
sir <- fortify_EpiModel(EpiModel_icm)
head(sir)
class(sir)
}
