% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify_data.R
\name{fortify_EpiModel}
\alias{fortify_EpiModel}
\title{Takes in output from the \code{R} \code{EpiModel} package and puts it in SIR format}
\usage{
fortify_EpiModel(EpiModel_output)
}
\arguments{
\item{EpiModel_output}{output from either \code{EpiModel::dcm} or \code{EpiModel::icm}}
}
\value{
data frame with the following columns
\describe{
\item{t}{the time}
\item{S}{number of Susceptibles}
\item{I}{number of Infectious}
\item{R}{number of Recovered}
\item{sim}{simulation number (factor variable) (optional column)}
}
}
\description{
Takes in output from the \code{R} \code{EpiModel} package and puts it in SIR format
}
\details{
Take the output from \code{EpiModel::dcm} and turn it into an SIR data.frame for plotting.
}
\examples{
## For dcm
sir1 <- fortify_EpiModel(EpiModel_det)
head(sir1)

## For icm
sir2 <- fortify_EpiModel(EpiModel_icm)
head(sir2)
}
