% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragment_info.R
\name{fragment_info}
\alias{fragment_info}
\title{Summary on fragments}
\usage{
fragment_info(picard_list, namelist)
}
\arguments{
\item{picard_list}{List of duplication metrics generated by Picard in data frame.
Data frames must be listed using `list()` even for one sample (see example below).
To import Picard duplication metrics (.txt file) into R, use `picard <- read.table("/path/to/picard/output", header = TRUE, fill = TRUE)`}

\item{namelist}{List of names in the same order as list of Picard outputs. Add multiple by `c()`}
}
\value{
A table summarizing metrics on fragments.
}
\description{
This function outputs a summary on fragments using metrics generated by Picard.
Provides the number of mapped fragments, duplication rate and number of unique fragments.
}
\examples{
library(EpiCompare)
data(CnT_H3K27ac_picard) # load example picard output
fragment_info(picard_list = list(CnT_H3K27ac_picard),
              namelist = "CUT&Tag")

}
