% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragment_info.R
\name{fragment_info}
\alias{fragment_info}
\title{Summary on fragments}
\usage{
fragment_info(picard_list)
}
\arguments{
\item{picard_list}{Named list of duplication metrics generated by Picard as data frame.
Data frames must be listed using `list()`.
To import Picard duplication metrics (.txt file) into R as data frame,
use `picard <- read.table("/path/to/picard/output", header = TRUE, fill = TRUE)`.
The list must be named e.g. `names(picard_list) <- c("name1","name2")`}
}
\value{
A table summarizing metrics on fragments.
}
\description{
This function outputs a summary on fragments using metrics generated by Picard.
Provides the number of mapped fragments, duplication rate and number of unique fragments.
}
\examples{
library(EpiCompare)
data(CnT_H3K27ac_picard) # load example picard output
data(CnR_H3K27ac_picard) # load example picard output

# To import Picard duplication metrics (.txt file) into R as data frame
# CnT_H3K27ac_picard <- read.table("/path/to/picard/output.txt", header = TRUE, fill = TRUE)

picard <- list(CnT_H3K27ac_picard, CnR_H3K27ac_picard) # convert into list
names(picard) <- c("CnT_H3K27ac", "CnR_H3K27ac") # set names
fragment_info(picard_list = picard)

}
