% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hagelloch_raw}
\alias{hagelloch_raw}
\title{Measles in Hagelloch, Germany, 1861}
\format{A data frame with 188 rows and 12 columns
\describe{
  \item{PN}{Patient number (Case ID) per individual}
  \item{NAME}{Patient's last name}
  \item{FN}{Family number (individuals with matching \code{FN} are in the
  same family)}
  \item{HN}{Household number ((individuals with matching \code{HN} are in the
  same household)}
  \item{AGE}{Age in years (fractions ignored)}
  \item{SEX}{Gender of the individual (factor: female; male)}
  \item{PRO}{\code{Date} of onset of prodromal symptoms}
  \item{ERU}{\code{Date} of onset of rash (not 100 sure of this one)}
  \item{CL}{School class (factor: preschool; 1st class; 2nd class )}
  \item{DEAD}{\code{Date} of death (with missings)}
  \item{IFTO}{number of patient who is the putative source of infection (0 = unknown)}
  \item{SI}{serial interval = number of days between dates of prodromes of infection source and infected person}
  \item{C}{Complications (factor: "no complications", or what complication)}
  \item{PR}{duration of prodromes in days}
  \item{CA}{number of cases in family}
  \item{NI}{number of initial cases}
  \item{GE}{generation number of the case}
  \item{TD}{day of max. fever (days after rush)}
  \item{TM}{max.fever (degree Celsius)}
  \item{x.loc}{x coordinate of house (in metres). Scaling in metres
  is obtained by multiplying the original coordinates by 2.5 (see details
  in Neal and Roberts (2004))}
  \item{y.loc}{y coordinate of house (in metres). See \code{x.loc} above}
  \item{tPRO}{exact time of onset of prodromal symptoms (relative to start
  of outbreak)}
  \item{tERU}{exact time of onset of rash (relative to start
  of outbreak) }
  \item{tDEAD}{time of death (relative to start of outbreak) (\code{NA}
  implies recovered)}
  \item{tR}{imputed time when individual entered into recovery}
  \item{tI}{imputed time when individual was infected}
}}
\source{
\url{https://rdrr.io/rforge/surveillance/man/hagelloch.html}
}
\usage{
hagelloch_raw
}
\description{
These data comprise of 188 cases of measles among children in the German
city of Hagelloch, 1861. The data were
originally collected by Dr. Albert Pfeilsticker (1863) and augmented and
re-analysed by Dr. Heike Oesterle (1992). This data comes from the package
\code{\link[outbreaks:measles_hagelloch_1861]{outbreaks}}.
}
\examples{
## show first few cases
head(hagelloch_raw)
}
\keyword{datasets}
