% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hg38_blacklist}
\alias{hg38_blacklist}
\title{Human genome hg38 blacklisted regions}
\format{
An object of class \code{GRanges} of length 910.
}
\source{
The code to prepare the .Rda file is:

\code{## blacklisted regions were directly downloaded} \cr
\code{## from https://www.encodeproject.org/files/ENCFF356LFX/} \cr
\code{hg38_blacklist <- ChIPseeker::readPeakFile(file.path(path), as = "GRanges")} \cr
\code{usethis::use_data(hg38_blacklist, overwrite = TRUE)} \cr
}
\usage{
data("hg38_blacklist")
}
\description{
Obtained from \url{https://www.encodeproject.org/files/ENCFF356LFX/}.
The ENCODE blacklist includes regions of the genome that have anomalous and/
or unstructured signals independent of the cell-line or experiment. Removal
of ENCODE blacklist is recommended for quality measure.
}
\keyword{datasets}
