% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liftover_grlist.R
\name{liftover_grlist}
\alias{liftover_grlist}
\title{Liftover peak list}
\usage{
liftover_grlist(grlist, input_build, output_build = "hg19")
}
\arguments{
\item{grlist}{A named list of \link[GenomicRanges]{GRanges} object.}

\item{input_build}{The genome build of \code{grlist}.}

\item{output_build}{Desired genome build for
\code{grlist} to be lifted over to.}
}
\value{
Named list of lifted \link[GenomicRanges]{GRanges} objects.
}
\description{
Perform genome build liftover to on a named list of
 \link[GenomicRanges]{GRanges} objects.
}
