% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance-depth.R
\name{local_distance_depth_function}
\alias{local_distance_depth_function}
\title{Localized version of Geenens & Nieto-Reyes functional distance-based depth}
\usage{
local_distance_depth_function(dist_mat, tau = Inf)
}
\arguments{
\item{dist_mat}{a n x n square positive symmetric matrix}

\item{tau}{localizing parameter (default is Inf)}
}
\value{
depth vector length n with depth values associated with indices in
  \code{dist_mat}
}
\description{
Calculates a localized distance-based depth vector using a distance matrix.
Specifically we alter Geenens & Nieto-Reyes's global distance-based depth to
a local depth similar to Agostinelli & Romanazzi (2011) localized approach,
and define our local distance-based depth as:
}
\details{
\eqn{LDD(x, \hat{P}, \tau) = 1/(|S| choose 2) \cdot \sum_{i!=j, i,j \in S}
I(d(X_i, X_j) > max(d(X_i,x), d(X_j,x)))}

where \eqn{S = {i : D(X_i,x) < \tau}}.

Note that if \eqn{|S| = 1}, then we say that \code{LDD(x) = 0}.


This function (renamed as \code{depth_function}) is shared with
\pkg{TCpredictionbands} on github:
\href{https://github.com/Mr8ND/TC-prediction-bands/tree/master/TCpredictionbands}{TCpredictionbands}.
}
\examples{
dist_mat <- matrix(c(0,   1, 1.5,
                     1,   0, 2,
                     1.5, 2, 0   ),
                   nrow = 3,
                   byrow = TRUE)

dd_vec <- local_distance_depth_function(dist_mat) # c(1,0,0)

ldd_vec1 <- local_distance_depth_function(dist_mat, tau = 2) # c(1,0,0)
ldd_vec2 <- local_distance_depth_function(dist_mat, tau = 1.5) # c(1,0,0)
ldd_vec3 <- local_distance_depth_function(dist_mat, tau = 1) # c(0,0,0)
ldd_vec <- local_distance_depth_function(dist_mat, tau = .1) # c(0,0,0)
}
