% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap_heatmap.R
\name{overlap_heatmap}
\alias{overlap_heatmap}
\title{Generate heatmap of percentage overlap}
\usage{
overlap_heatmap(peaklist, save_dir = NULL)
}
\arguments{
\item{peaklist}{A list of peak files as GRanges object.
Files must be listed using `list()` and named using `names()`
If not named, default file names will be assigned.}

\item{save_dir}{Name of file to save the interactive heatmap output as HTML}
}
\value{
An interactive heatmap
}
\description{
This function generates a heatmap showing percentage of overlapping peaks between peak files.
}
\examples{
library(EpiCompare)
data("encode_H3K27ac") # example dataset as GRanges object
data("CnT_H3K27ac") # example dataset as GRanges object

peaks <- list(encode_H3K27ac, CnT_H3K27ac) # create list
names(peaks) <- c("encode", "CnT") # set names

overlap_heatmap(peaklist = peaks)

}
