% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap_heatmap.R
\name{overlap_heatmap}
\alias{overlap_heatmap}
\title{Generate heatmap of percentage overlap}
\usage{
overlap_heatmap(peaklist, namelist)
}
\arguments{
\item{peaklist}{A list of peaks as GRanges object. Objects in list using `list()`}

\item{namelist}{A list of names. Names in list using `c()`}
}
\value{
An interactive heatmap
}
\description{
Generate heatmap of percentage overlap
}
