% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap_percent.R
\name{overlap_percent}
\alias{overlap_percent}
\title{Calculate percentage of overlapping peaks}
\usage{
overlap_percent(peaklist, reference, invert = FALSE)
}
\arguments{
\item{peaklist}{A list of peak files as GRanges object.
Files must be listed using `list()` and named using `names()`
If not named, default file names will be assigned.}

\item{reference}{refrence peaks file}

\item{invert}{whether to invert}
}
\value{
data frame
}
\description{
Calculate percentage of overlapping peaks
}
\examples{
data("encode_H3K27ac") # example dataset as GRanges object
data("CnT_H3K27ac") # example dataset as GRanges object
data("CnR_H3K27ac") # example dataset as GRanges object

peaks <- list(CnT_H3K27ac, CnR_H3K27ac) # create a list
names(peaks) <- c("CnT", "CnR") # set names
reference_peak <- list("ENCODE"=encode_H3K27ac)

overlap <- overlap_percent(peaklist=peaks,
                           reference=reference_peak[[1]])

}
