% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap_upset_plot.R
\name{overlap_upset_plot}
\alias{overlap_upset_plot}
\title{Generate Upset plot for overlapping peaks}
\usage{
overlap_upset_plot(peaklist)
}
\arguments{
\item{peaklist}{A named list of peak files as GRanges object.
Objects listed using `list()` and named using `names()`.
If not named, default file names are assigned.}
}
\value{
Upset plot of overlapping peaks
}
\description{
This function generates upset plot (UpSetR package) of overlapping peaks.
}
\examples{
data("encode_H3K27ac") # load example data
data("CnT_H3K27ac") # load example data
peakfile <- list(encode_H3K27ac, CnT_H3K27ac) # create list
names(peakfile) <- c("ENCODE","CnT") # name list
my_plot <- overlap_upset_plot(peaklist = peakfile) # run function
}
