% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_info.R
\name{peak_info}
\alias{peak_info}
\title{Summary of Peak Information}
\usage{
peak_info(peak_list, file_names, blacklist)
}
\arguments{
\item{peak_list}{A list of peak files as GRanges object. Objects in lists using `list()`.}

\item{file_names}{A list of file names in the correct order of peak_list. Names in list using `c()`.}

\item{blacklist}{A GRanges object containing blacklisted regions.}
}
\value{
A summary table of peak information
}
\description{
This function outputs a table summarizing information on the peak files.
Provides the total number of peaks and the percentage of peaks in blacklisted regions.
}
\examples{
library(EpiCompare)
data("encode_H3K27ac") # example dataset as GRanges object
data("CnT_H3K27ac") # example dataset as GRanges object
data("hg19_blacklist") # example blacklist dataset as GRanges object

peak_info(peak_list = list(encode_H3K27ac, CnT_H3K27ac),
          file_names = c("ENCODE", "CnT"),
          blacklist = hg19_blacklist)

}
