% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ChIPseeker_annotation.R
\name{plot_ChIPseeker_annotation}
\alias{plot_ChIPseeker_annotation}
\title{Create ChIPseeker annotation plot}
\usage{
plot_ChIPseeker_annotation(peaklist, annotation)
}
\arguments{
\item{peaklist}{A list of peak files as GRanges object.
Files must be listed using `list()` and named using `names()`
If not named, default file names will be assigned.}

\item{annotation}{A TxDb annotation object from Bioconductor.}
}
\value{
barplot
}
\description{
This function annotates peaks using `annotatePeak` from `ChIPseeker` package.
It outputs functional annotation of each peak file in a barplot.
}
\examples{
data("CnT_H3K27ac") # example dataset as GRanges object
data("CnR_H3K27ac") # example dataset as GRanges object

peaks <- list(CnT_H3K27ac, CnR_H3K27ac) # create a list
names(peaks) <- c("CnT", "CnR") # set names

## not run
# txdb<-TxDb.Hsapiens.UCSC.hg19.knownGene::TxDb.Hsapiens.UCSC.hg19.knownGene
# my_plot <- plot_ChIPseeker_annotation(peaklist = peaks
#                                       annotation = txdb)

}
