% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ChIPseeker_annotation.R
\name{plot_ChIPseeker_annotation}
\alias{plot_ChIPseeker_annotation}
\title{Create ChIPseeker annotation plot}
\usage{
plot_ChIPseeker_annotation(peaklist, txdb)
}
\arguments{
\item{peaklist}{A list of peak files as GRanges object.
Files must be listed using `list()` and named using `names()`
If not named, default file names will be assigned.}

\item{txdb}{A TxDb object.}
}
\value{
barplot
}
\description{
This function annotates peaks using `annotatePeak` from `ChIPseeker` package.
It uses annotation `TxDb.Hsapiens.UCSC.hg19.knownGene` provided by Bioconductor.
It outputs functional annotation of each peak file in a barplot.
}
