% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ChIPseeker_annotation.R
\name{plot_ChIPseeker_annotation}
\alias{plot_ChIPseeker_annotation}
\title{Create ChIPseeker annotation plot}
\usage{
plot_ChIPseeker_annotation(peaklist, namelist)
}
\arguments{
\item{peaklist}{A list of peak files as GRanges object. Files must be listed using `list()`}

\item{namelist}{A list of file names in the same order as the list of peak files. Use `c()` for multiple.}
}
\value{
barplot
}
\description{
This function annotates peaks using `annotatePeak` from `ChIPseeker` package.
It uses annotation `TxDb.Hsapiens.UCSC.hg19.knownGene` provided by Bioconductor.
It outputs functional annotation of each peak file in a barplot.
}
