% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_chrmHMM.R
\name{plot_chrmHMM}
\alias{plot_chrmHMM}
\title{Plot ChromHMM heatmap}
\usage{
plot_chrmHMM(peaklist, namelist, chrmHMM_annotation)
}
\arguments{
\item{peaklist}{A list of peaks as GRanges object. Objects in list using `list()`}

\item{namelist}{A list of names. Names in list using `c()`}

\item{chrmHMM_annotation}{ChromHMM annotation list}
}
\value{
ChromHMM heatmap
}
\description{
Creates a heatmap using outputs from ChromHMM using ggplot2.
The function takes a list of peakfiles, performs ChromHMM and outputs a heatmap.
ChromHMM annotation file must be loaded prior to using this function.
}
\examples{

}
