% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_chrmHMM.R
\name{plot_chrmHMM}
\alias{plot_chrmHMM}
\title{Plot ChromHMM heatmap}
\usage{
plot_chrmHMM(peaklist, chrmHMM_annotation, genome_build, interact = TRUE)
}
\arguments{
\item{peaklist}{A list of peak files as GRanges object.
Files must be listed using `list()` and named using `names()`
If not named, default file names will be assigned.}

\item{chrmHMM_annotation}{ChromHMM annotation list}

\item{genome_build}{The human genome reference build used to generate
peakfiles. "hg19" or "hg38".}

\item{interact}{Default TRUE. By default, the heatmaps are interactive.
If set FALSE, the function generates a static ChromHMM heatmap.}
}
\value{
ChromHMM heatmap
}
\description{
Creates a heatmap using outputs from ChromHMM using ggplot2.The function
takes a list of peakfiles, performs ChromHMM and outputs a heatmap. ChromHMM
annotation file must be loaded prior to using this function.
}
\examples{
data("CnT_H3K27ac") # example dataset as GRanges object
data("CnR_H3K27ac") # example dataset as GRanges object
data("chromHMM_annotation_K562") # example chromHMM annotation

peaks <- list(CnT_H3K27ac, CnR_H3K27ac) # create a list
names(peaks) <- c("CnT", "CnR") # set names

## not run
#my_plot <- plot_chrmHMM(peaklist=peaks,
#                        chrmHMM_annotation=chromHMM_annotation_K562,
#                        genome_build = "hg19")

}
