% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_enrichment.R
\name{plot_enrichment}
\alias{plot_enrichment}
\title{Generate enrichment analysis plots}
\usage{
plot_enrichment(peaklist, txdb, annotation)
}
\arguments{
\item{peaklist}{A list of peak files as GRanges object.
Files must be listed using `list()` and named using `names()`
If not named, default file names will be assigned.}

\item{txdb}{A TxDb object}

\item{annotation}{A genome wide annotation file}
}
\value{
KEGG and GO dot plots
}
\description{
This function runs KEGG and GO enrichment analysis of peak files
and generates dot plots.
}
