% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_enrichment.R
\name{plot_enrichment}
\alias{plot_enrichment}
\title{Generate enrichment analysis plots}
\usage{
plot_enrichment(peaklist, annotation)
}
\arguments{
\item{peaklist}{A list of peak files as GRanges object.
Files must be listed and named using \code{list()}.
e.g. \code{list("name1"=file1, "name2"=file2)}.
If not named, default file names will be assigned.}

\item{annotation}{A TxDb annotation object from Bioconductor.}
}
\value{
KEGG and GO dot plots
}
\description{
This function runs KEGG and GO enrichment analysis of peak files
and generates dot plots.
}
\examples{
### Load Data ###
data("CnT_H3K27ac") # example peakfile GRanges object
data("CnR_H3K27ac") # example peakfile GRanges object

### Create Named Peaklist ###
peaks <- list("CnT"=CnT_H3K27ac, "CnR"=CnR_H3K27ac)

## not run
# txdb<-TxDb.Hsapiens.UCSC.hg19.knownGene::TxDb.Hsapiens.UCSC.hg19.knownGene
# my_plot <- plot_enrichment(peaklist = peaks,
#                            annotation = txdb)

}
