% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_genome_builds.R
\name{prepare_genome_builds}
\alias{prepare_genome_builds}
\title{Prepare genome builds}
\usage{
prepare_genome_builds(genome_build)
}
\arguments{
\item{genome_build}{A named list indicating the human genome build used to 
generate each of the following inputs:
\itemize{
\item{"peakfiles" : }{Genome build for the \code{peakfiles} input.
Assumes genome build is the same for each element in the \code{peakfiles}
list.}
\item{"reference" : }{Genome build for the \code{reference} input.}
\item{"blacklist" : }{Genome build for the \code{blacklist} input.}
}
Example input list:\cr
 \code{genome_build = list(peakfiles="hg38", 
 reference="hg19", blacklist="hg19")}\cr\cr
Alternatively, you can supply a single character string instead of a list.
This should \emph{only} be done in situations where all three inputs
(\code{peakfiles}, \code{reference}, \code{blacklist}) are of the same 
genome build. For example:\cr
\code{genome_build = "hg19"}}
}
\value{
Named list.
}
\description{
Parse the \code{genome_build} argument into
\code{peaklist_build} and \code{reference_build}.
}
\keyword{internal}
