% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_peaklist.R
\name{prepare_peaklist}
\alias{prepare_peaklist}
\title{Prepare peaklist as GRanges}
\usage{
prepare_peaklist(peaklist)
}
\arguments{
\item{peaklist}{A named list of peaks as GRanges or paths to BED files}
}
\value{
A list of GRanges objects
}
\description{
Prepare peaklist as GRanges
}
\keyword{internal}
