% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bands.R
\docType{data}
\name{StatConfBandSpherical}
\alias{StatConfBandSpherical}
\title{stat object for use in delta_ball based stat_confidence_band and
geom_confidence_band}
\format{An object of class \code{StatConfBandDeltaBall} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 3.}
\usage{
StatConfBandSpherical
}
\description{
stat object for use in delta_ball based stat_confidence_band and
geom_confidence_band
}
\keyword{datasets}
