% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-dist-mat.R
\name{tidy_dist_mat}
\alias{tidy_dist_mat}
\title{creates tidy_dist_mat objects}
\usage{
tidy_dist_mat(dist_mat, rownames_df = NULL, colnames_df = NULL)
}
\arguments{
\item{dist_mat}{distance matrix}

\item{rownames_df}{data.frame with row identifying information}

\item{colnames_df}{data.frame with column identifying information}
}
\value{
\code{tidy_dist_mat} object
}
\description{
creates tidy_dist_mat objects
}
