% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tss_plot.R
\name{tss_plot}
\alias{tss_plot}
\title{Read count frequency around TSS}
\usage{
tss_plot(peaklist)
}
\arguments{
\item{peaklist}{A list of peak files as GRanges object.
Files must be listed using `list()` and named using `names()`
If not named, default file names will be assigned.}
}
\value{
profile plot
}
\description{
This function generates a plot of read count frequency around TSS.
}
\examples{
data("CnT_H3K27ac") # example dataset as GRanges object
data("CnR_H3K27ac") # example dataset as GRanges object
peaks <- list(CnT_H3K27ac, CnR_H3K27ac) # create a list
names(peaks) <- c("CnT", "CnR") # set names
#my_plot <- tss_plot(peaks)

}
