% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{UtoX_SIR_group}
\alias{UtoX_SIR_group}
\title{UtoX_SIR for grouped data frames}
\usage{
UtoX_SIR_group(U_g, T = NULL)
}
\arguments{
\item{U_g}{grouped data frame, with the following format
\describe{
  \item{init_state}{Initial state for individual (at time t = 0). For the
  states, 0 = S, 1 = I, 2 = R.}
  \item{max_time_S}{maximum time individual was suspectable (S)}
  \item{max_time_I}{maximum time individual was infected (I)}
}}

\item{T}{integer, max length of outbreak (default NULL)}
}
\value{
\code{sir_out} data frame, with columns
\describe{
  \item{grouping variable name(s)}{column/columns of grouping variable(s)}
  \item{t}{time since outbreak}
  \item{S}{Number of individuals suspectable}
  \item{I}{Number of individuals infected}
  \item{R}{Number of individuals in recovery}
}
}
\description{
UtoX_SIR for grouped data frames
}
\examples{
T <- 100
U_g <- hagelloch_raw \%>\% fortify_agents() \%>\% group_by(AGE2 = as.numeric(cut(AGE,3)))
sir_group <- UtoX_SIR_group(U_g, T)
U <- U_g \%>\%
  filter(AGE2 == 1) \%>\% ungroup()
sir_group1 <- UtoX_SIR(U, T)
sir_group_1 <- sir_group \%>\% filter(AGE2 == 1)
assertthat::are_equal(sir_group1,
                      sir_group_1 \%>\% select(t, S, I, R) \%>\% data.frame)
}
