% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/width_boxplot.R
\name{width_boxplot}
\alias{width_boxplot}
\title{Peak width boxplot}
\usage{
width_boxplot(peaklist, namelist)
}
\arguments{
\item{peaklist}{A list of peak files as GRanges object. Files must be listed using `list()`}

\item{namelist}{A list of file names in the same order as the list of peak files. Use `c()` for multiple.}
}
\value{
A boxplot of peak widths.
}
\description{
This function creates boxplots showing the distribution of widths in each peak file.
}
\examples{
library(EpiCompare)
data("encode_H3K27ac") # example dataset as GRanges object
data("CnT_H3K27ac") # example dataset as GRanges object

width_boxplot(peaklist = list(encode_H3K27ac, CnT_H3K27ac),
              namelist = c("ENCODE", "CnT"))

}
