% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/width_boxplot.R
\name{width_boxplot}
\alias{width_boxplot}
\title{Peak width boxplot}
\usage{
width_boxplot(peaklist)
}
\arguments{
\item{peaklist}{A list of peak files as GRanges object.
Files must be named and listed using \code{list()}.
e.g. \code{list("name1"=file1, "name2"=file2)}}
}
\value{
A boxplot of peak widths.
}
\description{
This function creates boxplots showing the distribution of widths in
each peak file.
}
\examples{
### Load Data ###
data("encode_H3K27ac") # example peaklist GRanges object
data("CnT_H3K27ac") # example peaklist GRanges object

### Create Named Peaklist ###
peaks <- list("encode"=encode_H3K27ac, "CnT"=CnT_H3K27ac)

### Run ###
my_plot <- width_boxplot(peaklist = peaks)

}
