% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/width_boxplot.R
\name{width_boxplot}
\alias{width_boxplot}
\title{Peak width boxplot}
\usage{
width_boxplot(peaklist)
}
\arguments{
\item{peaklist}{A list of peak files as GRanges object.
Files must be listed using `list()`.
Files must be named using `names(peaklist) <- c("sample1","sample2)`.
If not named, default file names will be assigned.}
}
\value{
A boxplot of peak widths.
}
\description{
This function creates boxplots showing the distribution of widths in
each peak file.
}
\examples{
data("encode_H3K27ac") # example dataset as GRanges object
data("CnT_H3K27ac") # example dataset as GRanges object

peaks <- list(encode_H3K27ac, CnT_H3K27ac) # create list
names(peaks) <- c("encode", "CnT") # set names

my_plot <- width_boxplot(peaklist = peaks)

}
